% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{add_clusters}
\alias{add_clusters}
\title{Add cluster number to a data frame}
\usage{
add_clusters(
  data,
  cols,
  newcol = NULL,
  k = 2,
  method = "kmeans",
  labels = TRUE,
  clean = TRUE
)
}
\arguments{
\item{data}{A dataframe.}

\item{cols}{A tidy selection of item columns.}

\item{newcol}{Name of the new cluster column as a character vector.
Set to NULL (default) to automatically build a name
from the common column prefix, prefixed with "cls_".}

\item{k}{Number of clusters to calculate.
Set to NULL to output a scree plot for up to 10 clusters
and automatically choose the number of clusters based on the elbow criterion.}

\item{method}{The method as character value. One of "kmeans" (default) or "pam".
For "kmeans" all items will be converted to numerical values and scaled using
\code{base::\link[base:scale]{scale}}. Euclidean distance is used.
For "pam", all items will be converted to categorical values by \link{data_cat}.
A Gower dissimilarity matrix is used.
Note that logical values are treated as asymmetrical, i.e. \code{FALSE} does not have a meaning, when computing the gower metric.
Therefore cases with only \code{FALSE} values (no annotations, no codes) are removed.}

\item{labels}{Whether to get the label of the cluster column from the common prefix of item column labels.}

\item{clean}{Prepare data by \link{data_clean}.}
}
\value{
The input tibble with an additional cluster column (factor, prefixed "cls_").
The fit result is stored in the attribute stats.cluster.fit, the item names in
stats.cluster.items, the scree-plot or silhouette-plot data in stats.cluster.diag and the method in
stats.cluster.method.
}
\description{
Clustering is either performed using \code{stats::\link[stats:kmeans]{kmeans}}
(method = "kmeans") on scaled numerical variables or using \code{cluster::\link[cluster:pam]{pam}} on a Gower
dissimilarity matrix computed by \code{cluster::\link[cluster:daisy]{daisy}}
(method = "pam").

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
library(volker)
ds <- volker::chatgpt

volker::add_clusters(ds, starts_with("cg_adoption"), k = 3)
volker::add_clusters(ds, starts_with("cg_adoption"), k = 3, method = "pam")
}
