% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{add_styles}
\alias{add_styles}
\title{Alias for report_styles}
\usage{
add_styles(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{add_styles} has been deprecated. Use \link{report_styles} instead.
}
\keyword{internal}
