% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{add_to_report}
\alias{add_to_report}
\title{Add an object to the report list or print it}
\usage{
add_to_report(obj, chunks)
}
\arguments{
\item{obj}{A volker table or plot}

\item{chunks}{The current report list}
}
\value{
A list with volker tables or plots
}
\description{
Add an object to the report list or print it
}
