% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{cfg_get_na_numbers}
\alias{cfg_get_na_numbers}
\title{Get configured na numbers}
\usage{
cfg_get_na_numbers(default = VLKR_NA_NUMBERS)
}
\arguments{
\item{default}{The default na numbers, if not explicitly provided by na.numbers or the options.
return A vector with numbers that should be treated as NAs}
}
\description{
Retrieves values either from the option or from the constant.
}
\keyword{internal}
