% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_check_has_column}
\alias{check_check_has_column}
\title{Check whether a column exist and stop if not}
\usage{
check_check_has_column(data, col, msg = NULL, stopit = T)
}
\arguments{
\item{data}{A data frame}

\item{col}{A column name}

\item{msg}{A custom error message if the check fails}

\item{stopit}{If true, the execution stops when a column is missing}
}
\value{
boolean Whether the column exists
}
\description{
Check whether a column exist and stop if not
}
