% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_is_categorical}
\alias{check_is_categorical}
\title{Check whether a column selection is categorical}
\usage{
check_is_categorical(data, cols, msg = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{A tidyselection of columns.}

\item{msg}{A custom error message if the check fails.}
}
\value{
boolean Whether the columns are categorical
}
\description{
Check whether a column selection is categorical
}
\keyword{internal}
