% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_is_dataframe}
\alias{check_is_dataframe}
\title{Check whether the object is a dataframe}
\usage{
check_is_dataframe(obj, msg = NULL, stopit = TRUE)
}
\arguments{
\item{obj}{The object to test.}

\item{msg}{Optional, a custom error message.}

\item{stopit}{Whether to stop execution with an error message.}
}
\value{
boolean Whether the object is a data.frame object.
}
\description{
Check whether the object is a dataframe
}
\keyword{internal}
