% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_method}
\alias{check_method}
\title{Check Validity of Correlation Method}
\usage{
check_method(method, stopit = TRUE, msg = NULL)
}
\arguments{
\item{method}{A character string specifying the correlation method to check.}

\item{stopit}{Whether to stop execution if method is invalid.}

\item{msg}{A custom error message if the check fails.}
}
\value{
logical whether method is valid.
}
\description{
Check Validity of Correlation Method
}
