% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_nonequal_columns}
\alias{check_nonequal_columns}
\title{Check whether there is no overlap between two column sets}
\usage{
check_nonequal_columns(data, cols, cross, msg = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{A tidyselection of columns.}

\item{cross}{A tidyselection of columns.}

\item{msg}{A custom error message if the check fails.}
}
\value{
boolean Whether the two column sets are different.
}
\description{
Check whether there is no overlap between two column sets
}
\keyword{internal}
