% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_columns}
\alias{clean_columns}
\title{Prepare metadata and types of dataframe columns}
\usage{
clean_columns(data)
}
\arguments{
\item{data}{Data frame}
}
\value{
Data frame
}
\description{
\itemize{
\item Removes the avector class from all columns
(comes from Sosci and prevents combining vectors)
}
}
