% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_tab}
\alias{cluster_tab}
\title{Get tables for clustering result}
\usage{
cluster_tab(
  data,
  cols,
  newcol = NULL,
  k = NULL,
  method = "kmeans",
  labels = TRUE,
  clean = TRUE,
  ...
)
}
\arguments{
\item{data}{A tibble.}

\item{cols}{A tidy selection of item columns or a single column with cluster values as a factor or logical.
If the column already contains a cluster result from \link{add_clusters}, it is used, and other parameters are ignored.
If no cluster result exists, it is calculated with \link{add_clusters}.}

\item{newcol}{Name of the new cluster column as a character vector.
Set to NULL (default) to automatically build a name
from the common column prefix, prefixed with "cls_".}

\item{k}{Number of clusters to calculate.
Set to NULL to automatically determine an optimal cluster number.
For kmeans, the number of clusters is automatically chosed
based on the elbow criterion after calculating within-sums of squares
for up to 10 clusters.
For pam, the number of clusters is automatically chosen based
on the maximum average silhouette for up to 10 clusters.}

\item{method}{The method as a character value, one of \code{kmeans} or \code{pam}. See \link{add_clusters} for further details.}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{tab_metrics}.}
}
\value{
A volker list with with three volker tabs: cluster centers, cluster counts, and clustering diagnostics.
}
\description{
Clustering is performed using \link{add_clusters}.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
library(volker)
data <- volker::chatgpt

cluster_tab(data, starts_with("cg_adoption"), k = 2)

}
\keyword{internal}
