% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{data_num}
\alias{data_num}
\title{Convert values to numeric values}
\usage{
data_num(data, cols)
}
\arguments{
\item{data}{A data frame containing the items to be converted.}

\item{cols}{A tidy selection of columns to convert.}
}
\value{
A data frame with the specified items reversed.
}
\description{
Convert values to numeric values
}
\keyword{internal}
