% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{data_onehot}
\alias{data_onehot}
\title{One-hot encode selected columns}
\usage{
data_onehot(data, ...)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{...}{Tidyselect expressions specifying columns to one-hot encode}
}
\value{
Data frame with one hot encoded data
}
\description{
One-hot encode selected columns
}
\keyword{internal}
