% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{data_rc_negatives}
\alias{data_rc_negatives}
\title{Recode negatives and output a warning}
\usage{
data_rc_negatives(data, cols)
}
\arguments{
\item{data}{Data frame}

\item{cols}{A tidy column selection}
}
\value{
Data frame
}
\description{
Recode negatives and output a warning
}
