% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{data_rev}
\alias{data_rev}
\title{Reverse the Values of Specified Items}
\usage{
data_rev(data, cols)
}
\arguments{
\item{data}{A data frame containing the items to be reversed.}

\item{cols}{A tidy selection of columns to reverse.
For example, if you want to calculate an index of the
two items "I feel bad about this" and "I like it",
both coded with 1=not at all to 5=fully agree,
you need to reverse one of them to make the codings compatible.}
}
\value{
A data frame with the specified items reversed.
}
\description{
Reverse the Values of Specified Items
}
\keyword{internal}
