% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{data_rm_empty}
\alias{data_rm_empty}
\title{Remove cases with only FALSE values in logical columns}
\usage{
data_rm_empty(data, cols)
}
\arguments{
\item{data}{Data frame.}

\item{cols}{A tidy column selection.}
}
\value{
Data frame without empty rows. Removal information is added to the misings attribute.
}
\description{
Remove cases with only FALSE values in logical columns
}
\keyword{internal}
