% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{data_rm_na_levels}
\alias{data_rm_na_levels}
\title{Remove NA levels}
\usage{
data_rm_na_levels(data, na.levels = TRUE)
}
\arguments{
\item{data}{Data frame}

\item{na.levels}{Residual values to remove from factor columns.
Either a character vector with residual values or TRUE to use defaults in \link{VLKR_NA_LEVELS}.
You can define default residual levels by setting the global option vlkr.na.levels
(e.g. \code{options(vlkr.na.levels=c("Not answered"))}).}
}
\value{
Data frame
}
\description{
Remove NA levels
}
\keyword{internal}
