% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{data_rm_na_numbers}
\alias{data_rm_na_numbers}
\title{Remove NA numbers}
\usage{
data_rm_na_numbers(data, na.numbers = TRUE)
}
\arguments{
\item{data}{Data frame}

\item{na.numbers}{Either a numeric vector with residual values or TRUE to use defaults in \link{VLKR_NA_NUMERIC}.
You can also define residual values by setting the global option vlkr.na.numbers
(e.g. \code{options(vlkr.na.numbers=c(-9))}).}
}
\value{
Data frame
}
\description{
Remove NA numbers
}
\keyword{internal}
