% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{data_rm_negatives}
\alias{data_rm_negatives}
\title{Remove negatives and output a warning}
\usage{
data_rm_negatives(data, cols)
}
\arguments{
\item{data}{Data frame}

\item{cols}{A tidy column selection}
}
\value{
Data frame
}
\description{
Remove negatives and output a warning
}
\keyword{internal}
