% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{data_rm_zeros}
\alias{data_rm_zeros}
\title{Remove zero values, drop missings and output a message}
\usage{
data_rm_zeros(data, cols)
}
\arguments{
\item{data}{Data frame.}

\item{cols}{A tidy column selection.}
}
\value{
Data frame.
}
\description{
Remove zero values, drop missings and output a message
}
\keyword{internal}
