% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{diagnostics_cooksd}
\alias{diagnostics_cooksd}
\title{Cook's distance plot}
\usage{
diagnostics_cooksd(fit)
}
\arguments{
\item{fit}{The lm fit object}
}
\value{
A ggplot object
}
\description{
Cook's distance plot
}
\examples{
library(volker)
data <- filter(volker::chatgpt, sd_gender != "diverse")

data <- add_model(data, use_work, metric = sd_age)

fit <- attr(data$use_work, "lm.fit")
diagnostics_cooksd(fit)

}
\keyword{internal}
