% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{diagnostics_qq}
\alias{diagnostics_qq}
\title{Normal Q-Q}
\usage{
diagnostics_qq(fit)
}
\arguments{
\item{fit}{The lm fit object}
}
\value{
A ggplot object
}
\description{
Normal Q-Q
}
\examples{
library(volker)
data <- filter(volker::chatgpt, sd_gender != "diverse")

data <- add_model(data, use_work, metric = sd_age)

fit <- attr(data$prd_use_work, "lm.fit")
diagnostics_qq(fit)

}
\keyword{internal}
