% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{diagnostics_resid_fitted}
\alias{diagnostics_resid_fitted}
\title{Residuals vs Fitted plot}
\usage{
diagnostics_resid_fitted(fit)
}
\arguments{
\item{fit}{The lm fit object}
}
\value{
A ggplot object
}
\description{
Residuals vs Fitted plot
}
\examples{
library(volker)
data <- filter(volker::chatgpt, sd_gender != "diverse")

data <- add_model(data, use_work, metric = sd_age)

fit <- attr(data$prd_use_work, "lm.fit")
diagnostics_resid_fitted(fit)
}
\keyword{internal}
