% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{diagnostics_scale_location}
\alias{diagnostics_scale_location}
\title{Scale-Location (Spread-Location)}
\usage{
diagnostics_scale_location(fit)
}
\arguments{
\item{fit}{The lm fit object}
}
\value{
A ggplot object
}
\description{
Scale-Location (Spread-Location)
}
\examples{
library(volker)
data <- filter(volker::chatgpt, sd_gender != "diverse")

data <- add_model(data, use_work, metric = sd_age)

fit <- attr(data$use_work, "lm.fit)
diagnostics_scale_location(fit")

}
\keyword{internal}
