% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.add_plot_attributes}
\alias{.add_plot_attributes}
\title{Add the volker classes and options}
\usage{
.add_plot_attributes(pl, rows = NULL)
}
\arguments{
\item{pl}{A ggplot2 object}
}
\value{
The plot
}
\description{
Add the volker classes and options
}
