% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{.add_to_vlkr_rprt}
\alias{.add_to_vlkr_rprt}
\title{Add an object to the report list}
\usage{
.add_to_vlkr_rprt(obj, chunks, tab = NULL)
}
\arguments{
\item{obj}{A new chunk (volker table, volker plot or character value)}

\item{chunks}{The current report list}

\item{tab}{A tabsheet name or NULL}
}
\value{
A volker report object
}
\description{
Add an object to the report list
}
\keyword{internal}
