% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree.R
\name{.agree_classification}
\alias{.agree_classification}
\title{Calculate classification performance indicators such as precision and recall.}
\usage{
.agree_classification(x, y, ids = NULL, category = NULL)
}
\arguments{
\item{x}{Vactor with values.}

\item{y}{Vector with sources, the first value is taken as ground truth source.
To change the ground truth, use a factor and set the first factor level to the appropriate value.}

\item{ids}{Vector of  Case IDs.}

\item{category}{If no category is provided, macro statistics are returned (along with the number of categories in the output).
Provide a category to get the statistics for this category only.
If values are boolean (TRUE / FALSE) and no category is provided, the category is always assumed to be "TRUE".}
}
\value{
A list with classification performance indicators.
}
\description{
Calculate classification performance indicators such as precision and recall.
}
\keyword{internal}
