% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree.R
\name{.agree_confusion}
\alias{.agree_confusion}
\title{Generate confusion matrix}
\usage{
.agree_confusion(
  data,
  cols,
  coders,
  ids = NULL,
  category = NULL,
  labels = TRUE
)
}
\arguments{
\item{data}{A tibble.}

\item{cols}{The columns holding codings.}

\item{coders}{The column holding coders.}

\item{ids}{The column holding case identifiers.}

\item{category}{Focus category or null.}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}
}
\value{
A tibble representing a confusion matrix.
}
\description{
Generate confusion matrix
}
\keyword{internal}
