% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree.R
\name{.agree_one}
\alias{.agree_one}
\title{Calculate agreement scores for one item}
\usage{
.agree_one(x, y, ids = NULL, category = NULL, method)
}
\arguments{
\item{x}{Codings.}

\item{y}{Coders.}

\item{ids}{Case IDs.}

\item{category}{If no category is provided, macro statistics are returned (along with the number of categories in the output).
Provide a category to get the statistics for this category only.
If values are boolean (TRUE / FALSE) and no category is provided, the category is always assumed to be "TRUE".}

\item{method}{One of "reliability" or "classification".}
}
\value{
A list with agreement coefficients.
For F1 calculations, macro statistics are returned (i.e. average of statistics for each category).
}
\description{
Calculate agreement scores for one item
}
\keyword{internal}
