% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree.R
\name{.agree_reliability}
\alias{.agree_reliability}
\title{Calculate reliability scores}
\usage{
.agree_reliability(x, y, ids = NULL)
}
\arguments{
\item{x}{Vector with codings.}

\item{y}{Vector with coders.}

\item{ids}{Vector with case IDs.}
}
\value{
A list with reliability coefficients.
}
\description{
Calculate reliability scores
}
\keyword{internal}
