% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.attr_insert}
\alias{.attr_insert}
\title{Insert a name-value-pair into an object attribute}
\usage{
.attr_insert(obj, key, name, value)
}
\arguments{
\item{obj}{The object.}

\item{key}{The attribute key.}

\item{name}{The name of a list item within the attribute.}

\item{value}{The value of the list item.}
}
\value{
The object with new attributes.
}
\description{
Insert a name-value-pair into an object attribute
}
\keyword{internal}
