% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.attr_transfer}
\alias{.attr_transfer}
\title{Transfer attributes from one to another object}
\usage{
.attr_transfer(to, from, keys)
}
\arguments{
\item{to}{The target object.}

\item{from}{The source object.}

\item{keys}{A character vector of attribute keys}
}
\value{
The target object with the updated attributes.
}
\description{
Transfer attributes from one to another object
}
\keyword{internal}
