% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{.cluster_dist_ss}
\alias{.cluster_dist_ss}
\title{Sum-of-squares decomposition for a dissimilarity matrix}
\usage{
.cluster_dist_ss(dissim, clustering)
}
\arguments{
\item{dissim}{A dist object.}

\item{clustering}{An integer vector with cluster assignments.}
}
\value{
A list with tot.withinss, betweenss and totss.
}
\description{
Uses the Huygens decomposition of total dispersion for an arbitrary
dissimilarity matrix, so that within- and between-cluster "sum of squares"
can be reported for distance-based methods such as PAM with Gower distance.
}
\keyword{internal}
