% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{.effect_npmi}
\alias{.effect_npmi}
\title{Calculate nmpi}
\usage{
.effect_npmi(data, col, cross, smoothing = 0, adjust = "fdr", test = TRUE, ...)
}
\arguments{
\item{data}{A tibble.}

\item{col}{The column holding factor values.}

\item{cross}{The column to correlate.}

\item{smoothing}{Add pseudocount. Calculate the pseudocount based on the number of trials
to apply Laplace's rule of succession.}

\item{adjust}{Performing multiple significance tests inflates the alpha error.
Thus, p values need to be adjusted according to the number of tests.
Set a method supported by  \code{stats::\link[stats:p.adjust]{p.adjust}},
e.g. "fdr" (the default) or "bonferroni". Disable adjustment with "none".}

\item{test}{Boolean; whether to perform significance tests (default TRUE).}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{tab_counts}.}
}
\value{
A volker tibble.
}
\description{
Calculate nmpi
}
\keyword{internal}
