% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{.effects_correlations}
\alias{.effects_correlations}
\title{Test whether correlations are different from zero}
\usage{
.effects_correlations(data, cols, cross, method = "p", labels = TRUE)
}
\arguments{
\item{data}{A tibble}

\item{cols}{The columns holding metric values}

\item{cross}{The columns holding metric values to correlate}

\item{method}{The output metrics, TRUE or p = Pearson's R, s = Spearman's rho}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}
}
\value{
A tibble with correlation results
}
\description{
Test whether correlations are different from zero
}
\keyword{internal}
