% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.factor_with_attr}
\alias{.factor_with_attr}
\title{Create a factor vector and preserve all attributes}
\usage{
.factor_with_attr(x, levels = NULL)
}
\arguments{
\item{x}{The source value, usually a character vector}

\item{levels}{The new levels}
}
\value{
A factor vector with the new levels
}
\description{
Create a factor vector and preserve all attributes
}
\keyword{internal}
