% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.get_density_mode}
\alias{.get_density_mode}
\title{Get the maximum density value in a density plot}
\usage{
.get_density_mode(data, col)
}
\arguments{
\item{data}{A tibble.}

\item{col}{A tidyselect column.}
}
\value{
The maximum density value.
}
\description{
Useful for placing geoms in the center of density plots
}
\keyword{internal}
