% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.get_fig_settings}
\alias{.get_fig_settings}
\title{Get plot size and resolution for the current output format from the config}
\usage{
.get_fig_settings()
}
\value{
A list with figure settings
}
\description{
Get plot size and resolution for the current output format from the config
}
\keyword{internal}
