% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.get_plot_limits}
\alias{.get_plot_limits}
\title{Guess plot limits from values}
\usage{
.get_plot_limits(values)
}
\arguments{
\item{values}{Numeric values vector}
}
\value{
Either a vector c(0,1) or c(-1,1) or the range of the values,, whatever covers the values better
}
\description{
Guess plot limits from values
}
\keyword{internal}
