% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.get_plot_resolution}
\alias{.get_plot_resolution}
\title{Get plot size and resolution for the current output format from the config}
\usage{
.get_plot_resolution()
}
\value{
A list with the elements
}
\description{
Get plot size and resolution for the current output format from the config
}
