% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.knit_plot}
\alias{.knit_plot}
\title{Knit volker plots}
\usage{
.knit_plot(pl)
}
\arguments{
\item{pl}{A ggplot object with vlkr_options.
The vlk_options are added by .to_vlkr_plot()
and provide information about the number of vertical items (rows)
and the maximum.}
}
\value{
Character string containing a html image tag, including the base64 encoded image.
}
\description{
Automatically calculates the plot height from
chunk options and volker options.
}
\details{
Presumptions:
\itemize{
\item a screen resolution of 72dpi
\item a default plot width of 7 inches = 504px
\item a default page width of 700px (vignette) or 910px (report)
\item an optimal bar height of 40px for 910px wide plots. i.e. a ratio of 0.04
\item an offset of one bar above and one bar below
}
}
\keyword{internal}
