% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{.knit_prepare}
\alias{.knit_prepare}
\title{Prepare markdown content for table rendering}
\usage{
.knit_prepare(x, wrap = FALSE)
}
\arguments{
\item{x}{Markdown text.}

\item{wrap}{Wrap text after the given number of characters.}
}
\value{
Markdown text with line breaks and escaped special characters.
}
\description{
Prepare markdown content for table rendering
}
\keyword{internal}
