% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{.knit_shorten}
\alias{.knit_shorten}
\title{Compact table printing with shortened names and values}
\usage{
.knit_shorten(df)
}
\arguments{
\item{df}{A data frame or tibble.}
}
\value{
A data fram with shortened column names and cell content.
}
\description{
Truncates long column names and long character values,
for more readable console output.
}
\details{
The default column name length is 30 and the cell values length is 40.
Override with \code{options(vlkr.trunc.columns=20)} and
\code{options(vlkr.trunc.cells=20)}.
}
\keyword{internal}
