% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim.R
\name{.outliers}
\alias{.outliers}
\title{Calculate outliers}
\usage{
.outliers(x, k = 1.5)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A list of outliers.
}
\description{
Calculate outliers
}
\keyword{internal}
