% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree.R
\name{.pair_agreement}
\alias{.pair_agreement}
\title{Helper: mean of pairwise agreements (nominal)}
\usage{
.pair_agreement(vals)
}
\arguments{
\item{vals}{A matrix with two columns for coders and ratings as values}
}
\value{
The share of agreement.
}
\description{
Helper: mean of pairwise agreements (nominal)
}
\keyword{internal}
