% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree.R
\name{.pair_disagreement}
\alias{.pair_disagreement}
\title{Helper: mean of pairwise disagreements (nominal)}
\usage{
.pair_disagreement(vals)
}
\arguments{
\item{vals}{A matrix with two columns for coders and ratings as values}
}
\value{
The share of disagreement.
}
\description{
Helper: mean of pairwise disagreements (nominal)
}
\keyword{internal}
