% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_add_attributes}
\alias{.plot_add_attributes}
\title{Add the volker classes and options}
\usage{
.plot_add_attributes(pl, rows = NULL, maxlab = NULL)
}
\arguments{
\item{pl}{A ggplot2 object}

\item{rows}{The number of items on the vertical axis}

\item{maxlab}{The character length of the longest label to be plotted
on the vertical axis}
}
\value{
The plot
}
\description{
Add the volker classes and options
}
