% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_bars}
\alias{.plot_bars}
\title{Helper function: plot grouped bar chart}
\usage{
.plot_bars(
  data,
  category = NULL,
  ci = FALSE,
  scale = NULL,
  limits = NULL,
  numbers = NULL,
  orientation = "horizontal",
  base = NULL,
  title = NULL
)
}
\arguments{
\item{data}{Data frame with the columns item, value, p, n and optionally w
If w is provided, the column width is generated according the w value, resulting in a mosaic plot.}

\item{category}{Optionally, a category to focus. All rows not matching the category will be filtered out.}

\item{ci}{Whether to plot error bars for 95\% confidence intervals. Provide the columns ci.low and ci.high in data.}

\item{scale}{Direction of the scale: 0 = no direction for categories,
-1 = descending or 1 = ascending values.}

\item{numbers}{Set to something that evaluates to TRUE and add the .values column to the data frame to ouput values on the bars.}

\item{orientation}{Whether to show bars (horizontal) or columns (vertical)}

\item{base}{The plot base as character or NULL.}

\item{title}{The plot title as character or NULL.}
}
\value{
A ggplot object.
}
\description{
Helper function: plot grouped bar chart
}
\keyword{internal}
