% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_cor}
\alias{.plot_cor}
\title{Helper function: plot cor and regression outputs}
\usage{
.plot_cor(
  data,
  ci = TRUE,
  base = NULL,
  limits = NULL,
  title = NULL,
  label = NULL
)
}
\arguments{
\item{data}{Dataframe with the columns item and value.
To plot errorbars, add the columns low and high and set the ci-paramater to TRUE.}

\item{ci}{Whether to plot confidence intervals. Provide the columns low and high in data.}

\item{base}{The plot base as character or NULL.}

\item{limits}{The scale limits.}

\item{title}{The plot title as character or NULL.}

\item{label}{The y axis label.}
}
\value{
A ggplot object.
}
\description{
Helper function: plot cor and regression outputs
}
\keyword{internal}
