% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_heatmap}
\alias{.plot_heatmap}
\title{Helper function: Heatmap}
\usage{
.plot_heatmap(
  data,
  values_col,
  numbers_col = NULL,
  base_n,
  title = TRUE,
  labels = TRUE
)
}
\arguments{
\item{data}{A tibble with item combinations in the first two columns.
Only if the item values are equal in both columns, titles are added to the axes.}

\item{values_col}{Name of the column containing correlation values, a character value.}

\item{numbers_col}{Name of the column containing values to plot on the tiles or NULL to hide numbers.}

\item{base_n}{The number of cases, plotted in the base line.}

\item{title}{If TRUE (default) shows a plot title derived from the column labels.
Disable the title with FALSE or provide a custom title as character value.}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}
}
\value{
A ggplot object
}
\description{
Helper function: Heatmap
}
\keyword{internal}
