% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_means}
\alias{.plot_means}
\title{Helper function to plot means as bars, e.g. for plot_item_metrics and plot_group_metrics}
\usage{
.plot_means(result, limits, numbers, .labels)
}
\arguments{
\item{result}{The result table of tab_item_metrics() or tab_group_metrics()}

\item{limits}{The scale limits. Set NULL to extract limits from the labels.}

\item{numbers}{The values to print on the bars: "m" or NULL}

\item{.labels}{If True (default) extracts item labels from the attributes, see get_labels()}
}
\value{
Plot
}
\description{
Helper function to plot means as bars, e.g. for plot_item_metrics and plot_group_metrics
}
