% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_scree}
\alias{.plot_scree}
\title{Helper function: scree plot}
\usage{
.plot_scree(data, k = NULL, lab_x = NULL, lab_y = NULL)
}
\arguments{
\item{data}{Dataframe with the factor or cluster number in the first column
and the metric in the second.}

\item{k}{Provide one of the values in the first column to color points up to this value.}

\item{lab_x}{Label of the x axis}

\item{lab_y}{Label of the y axis}
}
\value{
A vlkr_plot object
}
\description{
Helper function: scree plot
}
\keyword{internal}
