% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_silhouette}
\alias{.plot_silhouette}
\title{Helper function: silhouette plot}
\usage{
.plot_silhouette(data, k = NULL, lab_x = NULL, lab_y = NULL)
}
\arguments{
\item{data}{Dataframe with the number of clusters k in the first column
and the average silhouette width in the second.}

\item{k}{Provide one of the values in the first column to highlight the
point at this value (the selected number of clusters).}

\item{lab_x}{Label of the x axis}

\item{lab_y}{Label of the y axis}
}
\value{
A vlkr_plot object
}
\description{
Helper function: silhouette plot
}
\keyword{internal}
