% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_summary}
\alias{.plot_summary}
\title{Helper function: plot grouped line chart by summarising values}
\usage{
.plot_summary(
  data,
  ci = FALSE,
  scale = NULL,
  base = NULL,
  box = FALSE,
  limits = NULL,
  title = NULL
)
}
\arguments{
\item{data}{Dataframe with the columns item, value.}

\item{ci}{Whether to plot confidence intervals of the means.}

\item{scale}{Passed to the label scale function.}

\item{base}{The plot base as character or NULL.}

\item{box}{Whether to add boxplots.}

\item{title}{The plot title as character or NULL.}
}
\value{
A ggplot object.
}
\description{
Helper function: plot grouped line chart by summarising values
}
\keyword{internal}
