% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree.R
\name{.replace_first_col_labels}
\alias{.replace_first_col_labels}
\title{Helper: Remove common prefix from the first column}
\usage{
.replace_first_col_labels(result)
}
\arguments{
\item{result}{A tibble with item names in the first column}
}
\value{
A tibble with the first column renamed to the prefix and the prefix removed from column values.
}
\description{
Helper: Remove common prefix from the first column
}
\keyword{internal}
