% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{.report_agr}
\alias{.report_agr}
\title{Generate table and plot for agreement analysis}
\usage{
.report_agr(data, cols, coders, ids, method = "reliability", clean = TRUE, ...)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{Columns with codings. A tidy column selection,
e.g. a single column (without quotes)
or multiple columns selected by methods such as starts_with().}

\item{coders}{The coders or classification methods.}

\item{ids}{Column with case IDs.}

\item{method}{One of "reliability" or "classification".}
}
\value{
A list containing a table and a plot volker report chunk.
}
\description{
Generate table and plot for agreement analysis
}
\keyword{internal}
