% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{.report_fct}
\alias{.report_fct}
\title{Generate an factor table and plot}
\usage{
.report_fct(
  data,
  cols,
  cross,
  metric = FALSE,
  ...,
  k = 2,
  effect = FALSE,
  title = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{A tidy column selection,
e.g. a single column (without quotes)
or multiple columns selected by methods such as starts_with().}

\item{cross}{Not yet implemented. Optional, a grouping column (without quotes).}

\item{metric}{Not yet implemented. When crossing variables, the cross column parameter can contain categorical or metric values.
By default, the cross column selection is treated as categorical data.
Set metric to TRUE, to treat it as metric and calculate correlations.}

\item{k}{Number of factors to calculate.}

\item{effect}{Not yet implemented. Whether to report statistical tests and effect sizes.}

\item{title}{Add a plot title (default = TRUE).}
}
\value{
A list containing a table and a plot volker report chunk.
}
\description{
Generate an factor table and plot
}
\keyword{internal}
