% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{.report_idx}
\alias{.report_idx}
\title{Generate an index table and plot}
\usage{
.report_idx(data, cols, col_group, title = T)
}
\arguments{
\item{data}{A data frame}

\item{cols}{A tidy column selection,
e.g. a single column (without quotes)
or multiple columns selected by methods such as starts_with().}

\item{col_group}{Optional, a grouping column (without quotes).}

\item{title}{Add a plot title (default = TRUE)}
}
\value{
A list containing a table and a plot
}
\description{
Generate an index table and plot
}
\keyword{internal}
