% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{.report_mdl}
\alias{.report_mdl}
\title{Generate an model table and plot}
\usage{
.report_mdl(data, cols, categorical = NULL, metric = NULL, ..., title = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{A tidy column selection,
e.g. a single column (without quotes)
or multiple columns selected by methods such as starts_with().}

\item{metric}{Not yet implemented. When crossing variables, the cross column parameter can contain categorical or metric values.
By default, the cross column selection is treated as categorical data.
Set metric to TRUE, to treat it as metric and calculate correlations.}

\item{title}{Add a plot title (default = TRUE).}

\item{cross}{Not yet implementedt. Optional, a grouping column (without quotes).}

\item{k}{Number of factors to calculate.}

\item{effect}{Not yet implemented. Whether to report statistical tests and effect sizes.}
}
\value{
A list containing a table and a plot volker report chunk.
}
\description{
Generate an model table and plot
}
\keyword{internal}
