% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{.tab_split}
\alias{.tab_split}
\title{Split a metric column into categorical based on median}
\usage{
.tab_split(data, col, labels = TRUE)
}
\arguments{
\item{data}{A data frame containing the column to be split.}

\item{col}{The column to split.}

\item{labels}{Logical; if \code{TRUE} (default), use custom labels for the split categories based
on the column title. If \code{FALSE}, use the column name directly.}
}
\value{
A data frame with the specified column converted into categorical labels based on its median value.
The split threshold (median) is stored as an attribute of the column.
}
\description{
Split a metric column into categorical based on median
}
\keyword{internal}
