% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{.to_vlkr_df}
\alias{.to_vlkr_df}
\title{Add vlkr_df class - that means, the data frame has been prepared}
\usage{
.to_vlkr_df(data, digits = NULL)
}
\arguments{
\item{data}{A tibble}
}
\value{
A tibble of class vlkr_df
}
\description{
Add vlkr_df class - that means, the data frame has been prepared
}
\keyword{internal}
