% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{.to_vlkr_list}
\alias{.to_vlkr_list}
\title{Add vlkr_list class}
\usage{
.to_vlkr_list(data)
}
\arguments{
\item{data}{A list}
}
\value{
A volker list
}
\description{
Used to collect multiple tables in a list,
e.g. from regression outputs
}
\keyword{internal}
