% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.to_vlkr_plot}
\alias{.to_vlkr_plot}
\title{Add the volker class and options}
\usage{
.to_vlkr_plot(pl, rows = NULL, maxlab = NULL)
}
\arguments{
\item{pl}{A ggplot2 object}

\item{rows}{The number of items on the vertical axis. Will be automatically determined when NULL.}

\item{maxlab}{The character length of the longest label to be plotted. Will be automatically determined when NULL.
on the vertical axis}
}
\value{
The plot
}
\description{
Add the volker class and options
}
\keyword{internal}
