% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{.to_vlkr_rprt}
\alias{.to_vlkr_rprt}
\title{Add the vlkr_rprt class to an object}
\usage{
.to_vlkr_rprt(chunks)
}
\arguments{
\item{chunks}{A list of character strings.}
}
\value{
A volker report object: List of character strings with the vlkr_rprt class
containing the parts of the report.
}
\description{
Adding the class makes sure the appropriate printing function
is applied in markdown reports.
}
\keyword{internal}
