% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim.R
\name{.whisker_lower}
\alias{.whisker_lower}
\title{Calculate lower whisker in a boxplot}
\usage{
.whisker_lower(x, k = 1.5)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The lower whisker value
}
\description{
Calculate lower whisker in a boxplot
}
\keyword{internal}
