% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim.R
\name{.whisker_upper}
\alias{.whisker_upper}
\title{Calculate upper whisker in a boxplot}
\usage{
.whisker_upper(x, k = 1.5)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The upper whisker value.
}
\description{
Calculate upper whisker in a boxplot
}
\keyword{internal}
