% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{effect_counts_items_grouped_items}
\alias{effect_counts_items_grouped_items}
\title{Effect size and test for comparing multiple variables by multiple grouping variables}
\usage{
effect_counts_items_grouped_items(data, cols, cross, clean = TRUE, ...)

effect_counts_items_grouped_items(data, cols, cross, clean = TRUE, ...)
}
\arguments{
\item{data}{A tibble containing item measures.}

\item{cols}{Tidyselect item variables (e.g. starts_with...).}

\item{cross}{The grouping items.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{effect_counts}.}
}
\value{
A volker tibble.

A volker tibble.
}
\description{
\strong{Not yet implemented. The future will come.}

\strong{Not yet implemented. The future will come.}
}
\keyword{internal}
