% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{effect_counts_one}
\alias{effect_counts_one}
\title{Test homogeneity of category shares}
\usage{
effect_counts_one(data, col, clean = TRUE, expected = NULL, ...)
}
\arguments{
\item{data}{A tibble.}

\item{col}{The column holding factor values.}

\item{clean}{Prepare data by \link{data_clean}}

\item{expected}{Numeric vector of expected proportions for each category in \code{col}.
If NULL, assumes a uniform distribution across categories.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{effect_counts}.}
}
\value{
A volker tibble.
}
\description{
TODO: @JJ remove expected param and only assuming uniform?
}
\examples{
library(volker)
data <- volker::chatgpt
data <- data \%>\%
   filter(sd_gender != "diverse")

effect_counts_one(data, sd_gender, expected = c(0.3, 0.7))

}
\keyword{internal}
