% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{effect_metrics_items_grouped}
\alias{effect_metrics_items_grouped}
\title{Output confidence intervals of group means, F-Statistics and effect size (Eta^2)}
\usage{
effect_metrics_items_grouped(data, cols, cross, clean = T, ...)
}
\arguments{
\item{data}{A tibble containing item measures.}

\item{cols}{Tidyselect item variables (e.g. starts_with...).}

\item{cross}{The column holding groups to compare.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{effect_metrics}.}
}
\value{
A volker tibble.
}
\description{
\strong{Not yet implemented. The future will come.}
}
\keyword{internal}
