% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{effect_metrics_one}
\alias{effect_metrics_one}
\title{Test whether the mean is different from zero}
\usage{
effect_metrics_one(data, col, clean = T, ...)
}
\arguments{
\item{data}{A tibble.}

\item{col}{The column holding metric values.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{effect_metrics}.}
}
\value{
A volker tibble.
}
\description{
\strong{Not yet implemented. The future will come.}
}
\keyword{internal}
