% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{effects_counts_items}
\alias{effects_counts_items}
\title{Test whether shares differ}
\usage{
effects_counts_items(data, cols, clean = TRUE, percent = TRUE, ...)
}
\arguments{
\item{data}{A tibble containing item measures}

\item{cols}{Tidyselect item variables (e.g. starts_with...)}

\item{clean}{Prepare data by \link{data_clean}.}

\item{percent}{Proportions are formatted as percent by default. Set to FALSE to get bare proportions.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{plot_counts}.}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}
}
\value{
A volker tibble
}
\description{
Test whether shares differ
}
\keyword{internal}
