% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{effects_counts_items_cor}
\alias{effects_counts_items_cor}
\title{Correlate the values in multiple items and output effect sizes and tests}
\usage{
effects_counts_items_cor(data, cols, cross, clean = T, ...)
}
\arguments{
\item{data}{A tibble containing item measures.}

\item{cols}{Tidyselect item variables (e.g. starts_with...).}

\item{cross}{The target columns or NULL to calculate correlations within the source columns.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{effects_counts}.}
}
\value{
A volker tibble.
}
\description{
Correlate the values in multiple items and output effect sizes and tests
}
\keyword{internal}
