% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{effects_counts_one_cor}
\alias{effects_counts_one_cor}
\title{Output test statistics and effect size for categories correlated with a metric column}
\usage{
effects_counts_one_cor(data, col, cross, clean = TRUE, ...)
}
\arguments{
\item{data}{A tibble}

\item{col}{The column holding factor values}

\item{cross}{The column holding metric values}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{effects_counts}.}
}
\value{
A volker tibble
}
\description{
Output test statistics and effect size for categories correlated with a metric column
}
\examples{
library(volker)
data <- volker::chatgpt

effects_counts_one_cor(data, adopter, sd_gender)

}
\keyword{internal}
