% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{effects_metrics_items_grouped}
\alias{effects_metrics_items_grouped}
\title{Output confidence intervals of group means, F-Statistics and effect size (Eta^2)}
\usage{
effects_metrics_items_grouped(data, cols, cross, clean = T, ...)
}
\arguments{
\item{data}{A tibble}

\item{cols}{Tidyselect item variables (e.g. starts_with...)}

\item{cross}{The column holding groups to compare}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder}
}
\value{
A volker tibble
}
\description{
Output confidence intervals of group means, F-Statistics and effect size (Eta^2)
}
\examples{
library(volker)
data <- volker::chatgpt

effects_metrics_items_grouped(data, starts_with("cg_adoption"), sd_gender)


}
\keyword{internal}
