% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{effects_metrics_one}
\alias{effects_metrics_one}
\title{Test whether the mean is different from zero}
\usage{
effects_metrics_one(data, col, clean = T, ...)
}
\arguments{
\item{data}{A tibble}

\item{col}{The column holding metric values}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder}
}
\value{
A volker tibble
}
\description{
Test whether the mean is different from zero
}
\keyword{internal}
