% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{effects_metrics_one_grouped}
\alias{effects_metrics_one_grouped}
\title{Output t-test results, a regression table with estimates and macro statistics}
\usage{
effects_metrics_one_grouped(
  data,
  col,
  cross,
  method = "lm",
  negative = FALSE,
  labels = TRUE,
  clean = TRUE,
  ...
)
}
\arguments{
\item{data}{A tibble.}

\item{col}{The column holding metric values.}

\item{cross}{The column holding groups to compare.}

\item{method}{A character vector of methods, e.g. c("t.test","lm").
Supported methods are t.test (only valid if the cross column contains two levels)
and lm (regression results).}

\item{negative}{If FALSE (default), negative values are recoded as missing values.}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{effects_metrics}.}
}
\value{
A volker list object containing volker tables with the requested statistics.
}
\description{
#TODO: remove tidycat from the package, implement own function
#TODO: Fix bug, why is p in model statistics not above stars?
#TODO: Remove NA from base level in regression table (and by this fix p=0 showing as NA for the Intercept)
#TODO: Add ci for R squared
#TODO: Implement a parameter to select from different outputs (don't show t-test by default)
}
\examples{
library(volker)
data <- volker::chatgpt

effects_metrics_one_grouped(data, sd_age, sd_gender)

}
\keyword{internal}
