% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx.R
\name{get_alpha}
\alias{get_alpha}
\title{Get number of items and Cronbach's alpha of a scale added by add_index()}
\usage{
get_alpha(data)
}
\arguments{
\item{data}{A data frame column.}
}
\value{
A named list with with the keys "items" and "alpha".
}
\description{
TODO: Rename to index_tab, return volker list as in factor_tab()
}
\keyword{internal}
