% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{get_betas}
\alias{get_betas}
\title{Calculate standardized betas}
\usage{
get_betas(fit)
}
\arguments{
\item{fit}{A model fitted with lm()}
}
\value{
A data frame with a row for each term
}
\description{
Calculate standardized betas
}
\keyword{internal}
