% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{get_ci}
\alias{get_ci}
\title{Calculate ci values to be used for error bars on a plot}
\usage{
get_ci(x, conf = 0.95)
}
\arguments{
\item{x}{A numeric vector.}

\item{conf}{The confidence level.}
}
\value{
A named list with values for y, ymin, and ymax.
}
\description{
Calculate ci values to be used for error bars on a plot
}
\keyword{internal}
