% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{get_cohensd}
\alias{get_cohensd}
\title{Get Cohen's d for unpaired samples}
\usage{
get_cohensd(values, groups, conf = 0.95, pooled = FALSE)
}
\arguments{
\item{values}{A numeric vector.}

\item{groups}{A vector indicating groups.}

\item{conf}{The confidence interval.}

\item{pooled}{Whether to pool variances.}
}
\value{
A list with the elements d (Cohen's d), ci.low and ci.high (its confidence interval)
}
\description{
TODO: NOT FINISHED YET
}
\keyword{internal}
