% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{get_col_label}
\alias{get_col_label}
\title{Get a column label}
\usage{
get_col_label(data, col)
}
\arguments{
\item{data}{Labeled data frame}

\item{col}{The column name}
}
\value{
A character value
}
\description{
TODO: REPLACE BY get_title()
}
