% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_dfm.R
\name{get_dfm}
\alias{get_dfm}
\title{Get a pruned dfm from tokens}
\usage{
get_dfm(msg_tokens)
}
\arguments{
\item{msg_tokens}{A quanteda tokens object}
}
\value{
A quanteda dfm
}
\description{
Get a pruned dfm from tokens
}
\keyword{internal}
