% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{get_direction}
\alias{get_direction}
\title{Detect whether a scale is a numeric sequence}
\usage{
get_direction(data, cols, extract = TRUE)
}
\arguments{
\item{data}{The dataframe.}

\item{cols}{The tidy selection.}

\item{extract}{Whether to extract numeric values from characters.}
}
\value{
0 = an undirected scale, -1 = descending values, 1 = ascending values.
}
\description{
From all values in the selected columns, the numbers are extracted.
If no numeric values can be found, returns 0.
Otherwise, if any positive values form an ascending sequence, returns -1.
In all other cases, returns 1.
}
\keyword{internal}
