% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{get_etasq}
\alias{get_etasq}
\title{Calculate Eta squared}
\usage{
get_etasq(fit)
}
\arguments{
\item{fit}{A model}
}
\value{
A data frame with at least the column Eta2
}
\description{
Calculate Eta squared
}
\keyword{internal}
