% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{get_gini}
\alias{get_gini}
\title{Calculate the Gini coefficient}
\usage{
get_gini(x)
}
\arguments{
\item{x}{A vector of counts or other values}
}
\value{
The gini coefficient
}
\description{
Calculate the Gini coefficient
}
\keyword{internal}
