% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_labels}
\alias{get_labels}
\title{Get variable labels from their comment attributes}
\usage{
get_labels(data, cols)
}
\arguments{
\item{data}{A tibble}
}
\value{
A tibble with the columns:
- item_group: First part of the column name, up to an underscore
- item_class: The last class value of an item (e.g. numeric, factor)
- item_name: The column name,
- item_label: The comment attribute of the column
- value_name: In case a column has multiple attributes, the attribute names
- value_label: In case a column has multiple attributes, the attribute values
}
\description{
Get variable labels from their comment attributes
}
