% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{get_limits}
\alias{get_limits}
\title{Get the numeric range from the labels}
\usage{
get_limits(data, cols, negative = F)
}
\arguments{
\item{data}{The labeled data frame}

\item{cols}{A tidy variable selection}

\item{negative}{Whether to include negative values}
}
\value{
A list or NULL
}
\description{
Get the numeric range from the labels
}
\keyword{internal}
