% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{get_npmi}
\alias{get_npmi}
\title{Calculate nmpi}
\usage{
get_npmi(
  data,
  col,
  cross,
  category = NULL,
  smoothing = 0,
  adjust = "fdr",
  test = TRUE,
  ...
)
}
\arguments{
\item{data}{A tibble.}

\item{col}{The column holding factor values.}

\item{cross}{The column to correlate.}

\item{category}{A vector of values to focus. If not null, all other values will be removed from the result.}

\item{smoothing}{Add pseudocount. Calculate the pseudocount based on the number of trials
to apply Laplace's rule of succession.}

\item{adjust}{Performing multiple significance tests inflates the alpha error.
Thus, p values need to be adjusted according to the number of tests.
Set a method supported by  \code{stats::\link[stats:p.adjust]{p.adjust}},
e.g. "fdr" (the default) or "bonferroni". Disable adjustment with FALSE.}

\item{test}{Boolean; whether to perform significance tests (default TRUE).}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{tab_counts}.}
}
\value{
A volker tibble.
}
\description{
Calculate nmpi
}
\keyword{internal}
