% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_scale}
\alias{get_scale}
\title{Detect whether a scale is a numeric sequence}
\usage{
get_scale(data, cols, categories)
}
\arguments{
\item{data}{The dataframe}

\item{cols}{The tidy selection}

\item{categories}{Category names}
}
\value{
0 = an undirected scale, -1 = descending values, 1 = ascending values
}
\description{
Detect whether a scale is a numeric sequence
}
