% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{get_stars}
\alias{get_stars}
\title{Get significance stars from p values}
\usage{
get_stars(x)
}
\arguments{
\item{x}{A vector of p values.}
}
\value{
A character vector with significance stars.
}
\description{
Get significance stars from p values
}
\keyword{internal}
