% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{get_title}
\alias{get_title}
\title{Get a common title for a column selection}
\usage{
get_title(data, cols, default = NULL)
}
\arguments{
\item{data}{A tibble.}

\item{cols}{A tidy column selection.}

\item{default}{A character string used in case not prefix is found}
}
\value{
A character string.
}
\description{
Get a common title for a column selection
}
\keyword{internal}
