% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_dfm.R
\name{get_tokens}
\alias{get_tokens}
\title{Get cleaned tokens}
\usage{
get_tokens(msg)
}
\arguments{
\item{msg}{A data frame with the column "doc" holding a document identifier
and "text" holding the text}
}
\value{
A quanteda tokens object
}
\description{
Get cleaned tokens
}
\keyword{internal}
