% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gini}
\alias{gini}
\title{Calculate the unbiased Gini coefficient}
\usage{
gini(x)
}
\arguments{
\item{x}{A numeric vector representing the data for which the Gini coefficient is to be calculated.}
}
\value{
The unbiased Gini coefficient for the input data.
}
\description{
This function computes the unbiased Gini coefficient for a given vector of data.
}
\details{
The Gini coefficient is a measure of statistical dispersion representing the inequality of a distribution.
The unbiased Gini coefficient is a corrected version that provides a less biased estimate, especially for small sample sizes.
}
\author{
Made with the help of ChatGPT
}
