% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_column}
\alias{has_column}
\title{Check whether a column exist and stop if not}
\usage{
has_column(data, col, stopit = T)
}
\arguments{
\item{data}{A data frame}

\item{col}{A column name}

\item{stopit}{If true, the execution stops when a column is missing}
}
\value{
boolean
}
\description{
Check whether a column exist and stop if not
}
