% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{html_report}
\alias{html_report}
\title{Volker style HTML document format}
\usage{
html_report(...)
}
\arguments{
\item{...}{Additional arguments passed to html_document}
}
\value{
R Markdown output format
}
\description{
Based on the standard theme, tweaks the pill navigation
to switch between tables and plots.
To use the format, in the header of your Markdown document,
set \code{output: volker::html_report}.
}
\examples{
\dontrun{
# Add `volker::html_report` to the output options of your Markdown document:
#
# ```
# ---
# title: "How to create reports?"
# output: volker::html_report
# ---
# ```
}
}
