% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx.R
\name{idx_alpha}
\alias{idx_alpha}
\title{Get number of items and Cronbach's alpha of a scale added by idx_add()}
\usage{
idx_alpha(data)
}
\arguments{
\item{data}{A data frame column}
}
\value{
A named list with with the keys "items" and "alpha"
}
\description{
Get number of items and Cronbach's alpha of a scale added by idx_add()
}
\keyword{internal}
