% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{knit_chunks}
\alias{knit_chunks}
\title{Knit volker chunks}
\usage{
knit_chunks(data, ...)
}
\arguments{
\item{data}{A volker chunks object}
}
\value{
Formatted  table produced by \link{kable}
}
\description{
Knit volker chunks
}
\keyword{internal}
