% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{knit_plot}
\alias{knit_plot}
\title{Knit plot with automatically calculated height}
\usage{
knit_plot(pl)
}
\arguments{
\item{pl}{A ggplot object}
}
\value{
Character string containing a html image tag, including the base64 encoded image
}
\description{
Knit plot with automatically calculated height
}
