% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{knit_print.vlkr_plt}
\alias{knit_print.vlkr_plt}
\title{Printing method for volker plots when knitting}
\usage{
\method{knit_print}{vlkr_plt}(x, ...)
}
\arguments{
\item{x}{The volker plot.}

\item{...}{Further parameters passed to print().}
}
\value{
Knitr asis output
}
\description{
Printing method for volker plots when knitting
}
\examples{
library(volker)
data <- volker::chatgpt

pl <- plot_metrics(data, sd_age)
print(pl)

}
\keyword{internal}
