% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{label_scale}
\alias{label_scale}
\title{Wrap labels in plot scales}
\usage{
label_scale(x, scale)
}
\arguments{
\item{x}{The label vector.}

\item{scale}{A named label vector to select elements that should be wrapped.
Prevents numbers from being wrapped.}
}
\value{
A vevtor of wrapped labels.
}
\description{
Wrap labels in plot scales
}
\keyword{internal}
