% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labs_clear}
\alias{labs_clear}
\title{Remove all comments from the selected columns}
\usage{
labs_clear(data, cols, labels = NULL)
}
\arguments{
\item{data}{A tibble.}

\item{cols}{Tidyselect columns.}

\item{labels}{The attributes to remove. NULL to remove all attributes except levels and class.}
}
\value{
A tibble with comments removed.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
library(volker)
volker::chatgpt |>
  labs_clear()
}
