% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labs_impute}
\alias{labs_impute}
\title{Add missing residual labels in numeric columns that have at least one labeled value}
\usage{
labs_impute(data)
}
\arguments{
\item{data}{A tibble}
}
\value{
A tibble with added value labels
}
\description{
Add missing residual labels in numeric columns that have at least one labeled value
}
\keyword{internal}
