% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labs_prefix}
\alias{labs_prefix}
\title{Remove common prefix from the first column}
\usage{
labs_prefix(result)
}
\arguments{
\item{result}{A tibble with item names in the first column}
}
\value{
A tibble with the first column renamed to the prefix and the prefix removed from column values.
}
\description{
Remove common prefix from the first column
}
\keyword{internal}
