% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_relevance.R
\name{lda_add_relevance}
\alias{lda_add_relevance}
\title{Add relevance}
\usage{
lda_add_relevance(data, fit, prefix = "tpc_", lambda = 0.6, seed = 1852)
}
\description{
Add relevance
}
\keyword{internal}
