% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_topics.R
\name{lda_add_topic}
\alias{lda_add_topic}
\title{Perform topic modeling on a text column and add the relevance scores of each topic to each row}
\usage{
lda_add_topic(
  data,
  col_text,
  prefix = "tpc_",
  k = NULL,
  lambda = 0.6,
  seed = 1852,
  plot = F
)
}
\arguments{
\item{data}{A tibble}

\item{col_text}{The column containing text}

\item{prefix}{A prefix added to the new topic columns}

\item{k}{The number of clusters / topics.
If NULL, the number will be determined by metrics from the ldatuning package()}

\item{lambda}{The lambda value for relevance calculation of terms and example documents.
The higher the lambda, the more specific terms come out.}

\item{seed}{A seed for the random number generator, makes results reproducible. You should change it!}

\item{plot}{Whether to plot ldatuning results}
}
\value{
A data frame with a new column for each topic
}
\description{
Perform topic modeling on a text column and add the relevance scores of each topic to each row
}
