% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_relevance.R
\name{lda_get_termrelevance}
\alias{lda_get_termrelevance}
\title{Get term relevance per topic}
\usage{
lda_get_termrelevance(fit, lambda = 0.6)
}
\arguments{
\item{fit}{The topic model}

\item{lambda}{The weighting factor between prevalence and distinctiveness}
}
\value{
A data frame with relevance metrics
}
\description{
Get term relevance per topic
}
\keyword{internal}
