% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_report.R
\name{lda_print_topdocs}
\alias{lda_print_topdocs}
\title{Helper function}
\usage{
lda_print_topdocs(data, col_topic, col_text, col_rank, n = 3, trunc = 1000)
}
\description{
Helper function
}
\keyword{internal}
