% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_report.R
\name{lda_report}
\alias{lda_report}
\title{Summarize a text column by topic modeling}
\usage{
lda_report(data, col_text, k = NULL, lambda = 0.6, seed = 1852)
}
\arguments{
\item{data}{A data frame}

\item{col_text}{The text column}

\item{k}{The number of clusters / topics.
If NULL, the number will be determined by metrics from the ldatuning package()}

\item{lambda}{The lambda value for relevance calculation of terms and example documents.
The higher the lambda, the more specific terms come out.}

\item{seed}{Topic modeling, for finding k and for the final solution as well as
multi dimensional scaling of the results are based on simulations.
Fix the random number generator seed to a number for reproducible results.}
}
\value{
A data frame with topic columns
}
\description{
Summarize a text column by topic modeling
}
