% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lm_pvalue}
\alias{lm_pvalue}
\title{Get the p value of an lm model}
\usage{
lm_pvalue(fit)
}
\arguments{
\item{fit}{The result of an lm() call}
}
\value{
The p value of the model
}
\description{
Get the p value of an lm model
}
\keyword{internal}
