% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{map_label}
\alias{map_label}
\title{Get a label for a key}
\usage{
map_label(key, options = list())
}
\arguments{
\item{key}{The key}

\item{options}{A list of labels with their keys}
}
\value{
The label from the options that matches the key
}
\description{
Get a label for a key
}
\keyword{internal}
