% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_plot}
\alias{model_plot}
\title{Plot regression coefficients}
\usage{
model_plot(
  data,
  col,
  categorical,
  metric,
  interactions = NULL,
  labels = TRUE,
  clean = TRUE,
  ...
)
}
\description{
Alias of \link{model_metrics_plot}.
In the future, logistic regression and a method parameter to choose between logistic and linear regression will be implemented.
}
