% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_tab}
\alias{model_tab}
\title{Output a regression table with estimates and macro statistics
for multiple categorical or metric independent variables}
\usage{
model_tab(
  data,
  col,
  categorical,
  metric,
  interactions = NULL,
  labels = TRUE,
  clean = TRUE,
  ...
)
}
\description{
Alias of \link{model_metrics_tab}.
In the future, logistic regression and a method parameter to choose between logistic and linear regression will be implemented.
}
