% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{named.to.list}
\alias{named.to.list}
\title{Convert a named vector to a list}
\usage{
named.to.list(x)
}
\arguments{
\item{x}{A named vector or a list}
}
\value{
Lists are returned as is. Vectors are converted to lists with names as list names.
}
\description{
Convert a named vector to a list
}
\keyword{internal}
