% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{pdf_report}
\alias{pdf_report}
\title{Volker style PDF document format}
\usage{
pdf_report(...)
}
\arguments{
\item{...}{Additional arguments passed to pdf_document.}
}
\value{
R Markdown output format.
}
\description{
Based on the standard theme, tweaks tex headers.
To use the format, in the header of your Markdown document,
set \code{output: volker::pdf_report}.
}
\examples{
\dontrun{
# Add `volker::pdf_report` to the output options of your Markdown document:
#
# ```
# ---
# title: "How to create reports?"
# output: volker::pdf_report
# ---
# ```
}
}
