% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_items.R
\name{plot_compare_items}
\alias{plot_compare_items}
\title{Plot items}
\usage{
plot_compare_items(data)
}
\arguments{
\item{data}{A tibble containing only the selected items}
}
\description{
Plot items
}
