% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_counts_items_cor}
\alias{plot_counts_items_cor}
\title{Correlate categorical items with a metric column}
\usage{
plot_counts_items_cor(data, cols, cross, clean = TRUE, ...)
}
\arguments{
\item{data}{A tibble containing item measures.}

\item{cols}{The item columns that hold the values to summarize.}

\item{cross}{The column holding metric values to correlate.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{plot_counts}.}
}
\value{
A ggplot object.
}
\description{
Correlate categorical items with a metric column
}
\keyword{internal}
