% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_counts_one_grouped}
\alias{plot_counts_one_grouped}
\title{Plot frequencies cross tabulated with a grouping column}
\usage{
plot_counts_one_grouped(
  data,
  col,
  col_group,
  category = NULL,
  ordered = NULL,
  missings = F,
  prop = "total",
  numbers = NULL,
  title = T,
  labels = T
)
}
\arguments{
\item{data}{A tibble}

\item{col}{The column holding factor values}

\item{col_group}{The column holding groups to compare}

\item{category}{The value FALSE will force to plot all categories.
A character value will focus a selected category.
When NULL, in case of boolean values, only the TRUE category is plotted.}

\item{ordered}{Values can be nominal (0) or ordered ascending (1) descending (-1).
By default (NULL), the ordering is automatically detected.
An appropriate color scale should be choosen depending on the ordering.
For unordered values, the default scale is used.
For ordered values, shades of the VLKR_FILLGRADIENT option are used.}

\item{missings}{Include missing values (default FALSE)}

\item{prop}{The basis of percent calculation: "total" (the default) or "rows".
To display column proportions, swap the first column with the grouping column.}

\item{numbers}{The numbers to print on the bars: "n" (frequency), "p" (percentage) or both.}

\item{title}{If TRUE (default) shows a plot title derived from the column labels.
Disable the title with FALSE or provide a custom title as character value.}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}
}
\description{
Note: only non-missing cases are used to calculate the percentage.
}
