% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_item_counts}
\alias{plot_item_counts}
\title{Output frequencies for multiple variables}
\usage{
plot_item_counts(data, cols, values = c("n", "p"), .labels = T)
}
\arguments{
\item{data}{A tibble containing item measures}

\item{cols}{Tidyselect item variables (e.g. starts_with...)}

\item{values}{The values to print on the bars: n (frequency) or p (percentage)}

\item{.labels}{If True (default) extracts item labels from the attributes, see get_labels()}
}
\description{
Output frequencies for multiple variables
}
