% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_metrics_one}
\alias{plot_metrics_one}
\title{Output a histogram for a single metric variable}
\usage{
plot_metrics_one(data, col, title = T, .labels = T)
}
\arguments{
\item{data}{A tibble}

\item{col}{The columns holding metric values}

\item{.labels}{If True (default) extracts item labels from the attributes, see get_labels()}
}
\description{
Output a histogram for a single metric variable
}
