% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plot_multi_means}
\alias{plot_multi_means}
\title{Alias for plot_metrics_items_grouped}
\usage{
plot_multi_means(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{plot_multi_means} has been deprecated. Use \link{plot_metrics_items_grouped} instead.
}
\keyword{internal}
