% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{prepare_scale}
\alias{prepare_scale}
\title{Prepare the scale attribute values}
\usage{
prepare_scale(data)
}
\arguments{
\item{data}{A tibble with a scale attribute.}
}
\value{
A named list or NULL.
}
\description{
Prepare the scale attribute values
}
\keyword{internal}
