% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{print.vlkr_plt}
\alias{print.vlkr_plt}
\alias{plot.vlkr_plt}
\title{Printing method for volker plots}
\usage{
\method{print}{vlkr_plt}(x, ...)

\method{plot}{vlkr_plt}(x, ...)
}
\arguments{
\item{x}{The volker plot}

\item{...}{Further parameters passed to print()}
}
\description{
Printing method for volker plots
}
\keyword{internal}
