% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_labels}
\alias{remove_labels}
\title{Remove all comments from the selected columns}
\usage{
remove_labels(data, cols, labels = NULL)
}
\arguments{
\item{data}{A tibble}

\item{cols}{Tidyselect columns}

\item{labels}{The attributes to remove. NULL to remove all attributes.}
}
\value{
A tibble with comments removed
}
\description{
Remove all comments from the selected columns
}
