% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labs_replace_names}
\alias{labs_replace_names}
\title{Replace values in the "item" column by labels found in the dataset}
\usage{
labs_replace_names(result, data, cols)
}
\arguments{
\item{result}{The table with a "item" column,
e.g. produced by tab_item_count()}

\item{data}{The labeled dataset}

\item{cols}{The tidyselect columns}
}
\description{
Replace values in the "item" column by labels found in the dataset
}
