% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Automatically detect the table type and output table and plot}
\usage{
report(data, scopes)
}
\arguments{
\item{data}{A tibble}

\item{scopes}{A list of column prefixes, e.g. "UM" or "WB.}
}
\description{
For each column prefix:
\itemize{
\item If it matches a single column, use tab_var_counts and plot_var_counts
\item If it matches multiple columns, use tab_items_counts and plot_item_counts
}
}
