% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_add}
\alias{report_add}
\title{Add an object to the report list or print it}
\usage{
report_add(obj, chunks, tab = NULL)
}
\arguments{
\item{obj}{A volker table or plot}

\item{chunks}{The current report list}

\item{tab}{A tabsheet name or NULL}
}
\value{
A list with volker tables or plots
}
\description{
Add an object to the report list or print it
}
