% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprectated.R, R/report.R
\name{add_styles}
\alias{add_styles}
\alias{report_styles}
\title{Alias for report_styles}
\usage{
add_styles(...)

report_styles()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{add_styles()} has been deprecated. Use \code{report_styles()} instead.

Call report_styles() in one of your first code chunks to let the magic happen.
Make sure the code chunk's include option is set to TRUE.
Outputs a style tag with css styles that modify the standard bootstrap
theme.
}
\details{
See the html_report() format for an alternative to include the styles.
}
\keyword{internal}
