% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{set_col_label}
\alias{set_col_label}
\title{Set column labels by their comment attribute}
\usage{
set_col_label(data, cols, labels)
}
\arguments{
\item{data}{A data frame}

\item{cols}{Tidyselect column names}

\item{labels}{The new labels}
}
\description{
Set column labels by their comment attribute
}
