% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_item_labels}
\alias{set_item_labels}
\title{Set variable labels by setting their comment attributes}
\usage{
set_item_labels(data, labels)
}
\arguments{
\item{data}{A tibble}

\item{labels}{A tibble with variable names in the first column (item_name) and their labels in the second column (item_label)}
}
\value{
A tibble with new variable labels
}
\description{
Set variable labels by setting their comment attributes
}
