% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim.R
\name{skim_boxplot}
\alias{skim_boxplot}
\title{A skimmer for boxplot generation}
\usage{
skim_boxplot(data, ..., .data_name = NULL)
}
\description{
Returns a five point summary, mean and sd, items count and alpha for scales added by idx_add().
Additionally, the whiskers defined by the minimum respective maximum value within 1.5 * iqr are calculated.
Outliers are returned in a list column.
}
\keyword{internal}
