% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim.R
\name{skim_grouped}
\alias{skim_grouped}
\title{Calculate a metric by groups}
\usage{
skim_grouped(data, cols, cross, value = "numeric.mean", labels = TRUE)
}
\arguments{
\item{data}{A tibble.}

\item{cols}{The item columns that hold the values to summarize.}

\item{cross}{The column holding groups to compare.}

\item{value}{The metric to extract from the skim result, e.g. numeric.mean or numeric.sd.}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}
}
\value{
A tibble with each item in a row, a total column and columns for all groups.
}
\description{
Calculate a metric by groups
}
\keyword{internal}
