% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stat_counts}
\alias{stat_counts}
\title{Output effect sizes for count data}
\usage{
stat_counts(data, cols, col_group = NULL, clean = TRUE, ...)
}
\arguments{
\item{data}{A data frame}

\item{cols}{A tidy column selection,
e.g. a single column (without quotes)
or multiple columns selected by methods such as starts_with()}

\item{col_group}{Optional, a grouping column. The column name without quotes.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Other parameters passed to the appropriate stat function}
}
\value{
A volker tibble
}
\description{
The type of effect size depends on the number of selected columns:
\itemize{
\item One column: see \link{stat_counts_one}
\item Multiple columns: see \link{stat_counts_items}
\item One column and one grouping column: see \link{stat_counts_one_grouped}
\item Multiple columns and one grouping column: see \link{stat_counts_items_grouped}
}
}
\examples{
library(volker)
data <- volker::chatgpt

stat_counts(data, sd_gender)

}
