% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stat_counts_items}
\alias{stat_counts_items}
\title{Output confidence intervals and effect size for multiple variables}
\usage{
stat_counts_items(data, cols, clean = TRUE, percent = TRUE, ...)
}
\arguments{
\item{data}{A tibble containing item measures}

\item{cols}{Tidyselect item variables (e.g. starts_with...)}

\item{clean}{Prepare data by \link{data_clean}.}

\item{percent}{Proportions are formatted as percent by default. Set to FALSE to get bare proportions.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{plot_counts}.}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}
}
\value{
A volker tibble
}
\description{
Output confidence intervals and effect size for multiple variables
}
\examples{
library(volker)
data <- volker::chatgpt

stat_counts_items(data, starts_with("cg_adoption_"))

}
\keyword{internal}
