% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stat_counts_items_cor}
\alias{stat_counts_items_cor}
\title{Correlate the values in multiple items and output effect sizes and tests}
\usage{
stat_counts_items_cor(data, cols, cross, clean = T, ...)
}
\arguments{
\item{data}{A tibble containing item measures}

\item{cols}{Tidyselect item variables (e.g. starts_with...)}

\item{cross}{The target columns or NULL to calculate correlations within the source columns}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder}
}
\value{
A volker tibble
}
\description{
Correlate the values in multiple items and output effect sizes and tests
}
\examples{
library(volker)
data <- volker::chatgpt

stat_counts_items_cor(ds, starts_with("cg_adoption_"), )

}
\keyword{internal}
