% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stat_counts_items_grouped}
\alias{stat_counts_items_grouped}
\title{Output confidence intervals and effect size for multiple variables by a grouping variable}
\usage{
stat_counts_items_grouped(data, cols, cross, clean = T, ...)
}
\arguments{
\item{data}{A tibble containing item measures and grouping variable}

\item{cols}{Tidyselect item variables (e.g. starts_with...)}

\item{cross}{The column holding groups to compare}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder}
}
\value{
A volker tibble
}
\description{
Output confidence intervals and effect size for multiple variables by a grouping variable
}
\examples{
library(volker)
data <- volker::chatgpt

stat_counts_items_grouped(data, starts_with("cg_adoption_"))

}
\keyword{internal}
