% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stat_counts_one}
\alias{stat_counts_one}
\title{Output confidence intervals for category shares}
\usage{
stat_counts_one(data, col, col_group, clean = TRUE, ...)
}
\arguments{
\item{data}{A tibble}

\item{col}{The column holding factor values}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{stat_counts}.}
}
\value{
A volker tibble
}
\description{
Output confidence intervals for category shares
}
\examples{
library(volker)
data <- volker::chatgpt

stat_counts_one(ds, adopter)

}
\keyword{internal}
