% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stat_counts_one_grouped}
\alias{stat_counts_one_grouped}
\title{Output test statistics and effect size for contingency tables (Chi^2 and Cramer's V)}
\usage{
stat_counts_one_grouped(data, col, col_group, clean = TRUE, ...)
}
\arguments{
\item{data}{A tibble}

\item{col}{The column holding factor values}

\item{col_group}{The column holding groups to compare}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{stat_counts}.}
}
\value{
A volker tibble
}
\description{
Output test statistics and effect size for contingency tables (Chi^2 and Cramer's V)
}
\examples{
library(volker)
data <- volker::chatgpt

stat_counts_one_grouped(ds, adopter, sd_gender)

}
\keyword{internal}
