% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stat_metrics_items}
\alias{stat_metrics_items}
\title{Output test statistics and effect size (Cohen's d) for paired samples}
\usage{
stat_metrics_items(data, cols, clean = T, ...)
}
\arguments{
\item{data}{A tibble}

\item{cols}{The column holding metric values}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder}
}
\value{
A volker tibble
}
\description{
Output test statistics and effect size (Cohen's d) for paired samples
}
\examples{
library(volker)
data <- volker::chatgpt

stat_metrics_items(data, starts_with("cg_adoption"))


}
\keyword{internal}
