% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stat_metrics_items_cor}
\alias{stat_metrics_items_cor}
\title{Output correlation coefficents for items}
\usage{
stat_metrics_items_cor(data, cols, clean = T, ...)
}
\arguments{
\item{data}{A tibble}

\item{cols}{The item columns that hold the values to summarize}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder}

\item{col_group}{The column holding groups to compare and test}
}
\value{
A volker tibble
}
\description{
Output correlation coefficents for items
}
\examples{
library(volker)
data <- volker::chatgpt

stat_metrics_items_grouped(data, starts_with("cg_adoption"), sd_gender)


}
\keyword{internal}
