% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stat_metrics_items_grouped}
\alias{stat_metrics_items_grouped}
\title{Output confindence intervals of group means, F-Statistics and effectsize (Eta^2)}
\usage{
stat_metrics_items_grouped(data, cols, cross, clean = T, ...)
}
\arguments{
\item{data}{A tibble}

\item{cols}{The item columns that hold the values to summarize}

\item{cross}{The column holding groups to compare and test}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder}
}
\value{
A volker tibble
}
\description{
Output confindence intervals of group means, F-Statistics and effectsize (Eta^2)
}
\examples{
library(volker)
data <- volker::chatgpt

stat_metrics_items_grouped(data, starts_with("cg_adoption"), sd_gender)


}
\keyword{internal}
