% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stat_metrics_one}
\alias{stat_metrics_one}
\title{Test whether the mean is different from zero}
\usage{
stat_metrics_one(data, col, clean = T, ...)
}
\arguments{
\item{data}{A tibble}

\item{col}{The column holding metric values}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder}
}
\value{
A volker tibble
}
\description{
Test whether the mean is different from zero
}
\examples{
library(volker)
data <- volker::chatgpt

stat_metrics_one(data, sd_age)

}
\keyword{internal}
