% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stat_metrics_one_grouped}
\alias{stat_metrics_one_grouped}
\title{Output t-test results, a regression table with estimates and macro statistics}
\usage{
stat_metrics_one_grouped(
  data,
  col,
  cross,
  negative = FALSE,
  labels = TRUE,
  clean = TRUE,
  ...
)
}
\arguments{
\item{data}{A tibble}

\item{col}{The column holding metric values}

\item{cross}{The column holding groups to compare}

\item{negative}{If FALSE (default), negative values are recoded as missing values.}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{stat_metrics}.}
}
\value{
A volker chunks object containing tables for micro and macro statistics
}
\description{
#TODO: Fix bug, why is p in model statistics not above stars?
#TODO: Remove NA from base level in regression table (and by this fix p=0 showing as NA for the Intercept)
#TODO: Add ci for R squared
#TODO: Implement a parameter to select from different outputs (don't show t-test by default)
}
\examples{
library(volker)
data <- volker::chatgpt

stat_metrics_one_grouped(data, sd_age, sd_gender)

}
\keyword{internal}
