% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_compare_factor}
\alias{tab_compare_factor}
\title{Compare & Table factor value by group}
\usage{
tab_compare_factor(data, col_category, col_group)
}
\arguments{
\item{data}{A tibble}

\item{col_category}{The column holding factor values}

\item{col_group}{The column holding groups to compare}
}
\description{
Compare & Table factor value by group
}
