% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_compare_items}
\alias{tab_compare_items}
\title{Compare & Table items}
\usage{
tab_compare_items(data, cols_items, col_group)
}
\arguments{
\item{data}{A tibble containing item measures}

\item{cols_items}{Tidyselect item variables (e.g. starts_with...)}

\item{col_group}{Optional faceting variable}
}
\description{
Compare & Table items
}
