% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_counts_items}
\alias{tab_counts_items}
\title{Output frequencies for multiple variables}
\usage{
tab_counts_items(
  data,
  cols,
  missings = F,
  values = c("n", "p"),
  percent = T,
  labels = T
)
}
\arguments{
\item{data}{A tibble containing item measures}

\item{cols}{Tidyselect item variables (e.g. starts_with...)}

\item{missings}{Include missing values (default FALSE)}

\item{values}{The values to output: n (frequency) or p (percentage) or both (the default)}

\item{percent}{Set to FALSE to prevent calculating percents from proportions}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{get_codebook}.}
}
\description{
TODO: Reorder boolean categories: first TRUE, then FALSE
TODO: Support single columns
}
