% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_counts_items_cor}
\alias{tab_counts_items_cor}
\title{Correlate the values in multiple items}
\usage{
tab_counts_items_cor(data, cols, cross, clean = TRUE, ...)
}
\arguments{
\item{data}{A tibble containing item measures.}

\item{cols}{The source columns.}

\item{cross}{The target columns or NULL to calculate correlations within the source columns.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{tab_counts}.}
}
\description{
\strong{Not yet implemented. The future will come.}
}
\keyword{internal}
