% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_counts_items_grouped}
\alias{tab_counts_items_grouped}
\title{Compare the values in multiple items by a grouping column.}
\usage{
tab_counts_items_grouped(
  data,
  cols,
  cols_groups,
  values = c("n", "p"),
  category = NULL,
  digits = 1,
  .labels = T
)
}
\arguments{
\item{data}{A tibble}

\item{cols}{The item columns that hold the values to summarize}

\item{values}{The output values, n or p or both}

\item{category}{The category within the items to focus}

\item{digits}{The digits to print}

\item{.labels}{If True (default) extracts item labels from the attributes, see get_labels()}

\item{cols_group}{The columns holding groups to compare}

\item{.negative}{If True (default), negative values are recoded to missing values}
}
\description{
TODO: implement -> focus one category and show n / p
}
\keyword{internal}
