% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_counts_items_grouped_items}
\alias{tab_counts_items_grouped_items}
\title{Correlation of categorical items with categorical items}
\usage{
tab_counts_items_grouped_items(
  data,
  cols,
  cross,
  method = "cramer",
  category = NULL,
  labels = TRUE,
  clean = TRUE,
  ...
)
}
\arguments{
\item{data}{A tibble containing item measures.}

\item{cols}{Tidyselect item variables (e.g. starts_with...).}

\item{cross}{Tidyselect item variables (e.g. starts_with...).}

\item{method}{The method of correlation calculation:
\itemize{
\item \code{cramer} for Cramer's V,
\item \code{npmi} for Normalized Pointwise Mutual Information. Not implemented yet.
}}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{plot_counts}.}
}
\value{
A volker tibble.
}
\description{
Correlation of categorical items with categorical items
}
\keyword{internal}
