% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_counts_one}
\alias{tab_counts_one}
\title{Output a frequency table for the values in one column}
\usage{
tab_counts_one(data, col, missings = F, percent = T, labels = T)
}
\arguments{
\item{data}{A tibble}

\item{col}{The column holding values to count}

\item{percent}{Proportions are formatted as percent by default. Set to FALSE to get bare proportions.}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{get_labels}.}
}
\description{
Output a frequency table for the values in one column
}
