% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_counts_one_cor}
\alias{tab_counts_one_cor}
\title{Correlate categorical groups with one metric column}
\usage{
tab_counts_one_cor(data, cols, cross, clean = TRUE, ...)
}
\arguments{
\item{data}{A tibble.}

\item{cols}{The item columns that hold the values to summarize.}

\item{cross}{The column to correlate.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{tab_counts}.}
}
\value{
A volker tibble.
}
\description{
\strong{Not yet implemented. The future will come.}
}
\keyword{internal}
