% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_counts_one_grouped}
\alias{tab_counts_one_grouped}
\title{Output frequencies cross tabulated with a grouping column}
\usage{
tab_counts_one_grouped(
  data,
  col,
  col_group,
  missings = F,
  prop = "total",
  values = c("n", "p"),
  percent = T,
  labels = T,
  ...
)
}
\arguments{
\item{data}{A tibble}

\item{col}{The column holding factor values}

\item{col_group}{The column holding groups to compare}

\item{missings}{Include missing values in the output (default FALSE)}

\item{prop}{The basis of percent calculation: "total" (the default), "cols", or "rows".}

\item{values}{The values to output: n (frequency) or p (percentage) or both (the default).}

\item{percent}{Proportions are formatted as percent by default. Set to FALSE to get bare proportions.}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}
}
\description{
Output frequencies cross tabulated with a grouping column
}
