% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{tab_counts_var}
\alias{tab_counts_var}
\title{Alias for tab_counts_one}
\usage{
tab_counts_var(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{tab_counts_var} has been deprecated. Use \link{tab_counts_one} instead.
}
\keyword{internal}
