% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_group_counts}
\alias{tab_group_counts}
\title{Count and compare groups (cross table)}
\usage{
tab_group_counts(data, col, col_group, value = "n", digits = 1, .quiet = F)
}
\arguments{
\item{data}{A tibble}

\item{col}{The column holding factor values}

\item{col_group}{The column holding groups to compare}

\item{value}{The value to output: n (frequency) or p (percentage)}

\item{digits}{The digits to print}

\item{.quiet}{Set to true to suppress printing the output}
}
\description{
Count and compare groups (cross table)
}
