% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{tab_group_metrics}
\alias{tab_group_metrics}
\title{Alias for tab_metrics_one_grouped}
\usage{
tab_group_metrics(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{tab_group_metrics} has been deprecated. Use \link{tab_metrics_one_grouped} instead.
}
\keyword{internal}
