% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_item_metrics}
\alias{tab_item_metrics}
\title{Output a five point summary table for multiple items}
\usage{
tab_item_metrics(data, cols, digits = 1, .quiet = F)
}
\arguments{
\item{data}{A tibble}

\item{digits}{The digits to print}

\item{.quiet}{Set to true to suppress printing the output}

\item{cols_values}{The columns holding metric values}
}
\description{
TODO: add labels (using get_labels())
}
