% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_means}
\alias{tab_means}
\title{Output the means for groups in one or multiple columns}
\usage{
tab_means(data, cols_groups, cols_values, digits = 1, .quiet = F)
}
\arguments{
\item{data}{A tibble}

\item{cols_values}{The columns that hold the values to calculate the mean}

\item{digits}{The digits to print}

\item{.quiet}{Set to true to suppress printing the output}

\item{cols_group}{The columns holding groups to compare}
}
\description{
Output the means for groups in one or multiple columns
}
