% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_metrics_items}
\alias{tab_metrics_items}
\title{Output a five point summary table for multiple items}
\usage{
tab_metrics_items(data, cols, digits = 1, .negative = F, .labels = T)
}
\arguments{
\item{data}{A tibble}

\item{cols}{The columns holding metric values}

\item{digits}{The digits to print}

\item{.negative}{If True (default), negative values are recoded to missing values}

\item{.labels}{If True (default) extracts item labels from the attributes, see get_labels()}
}
\description{
Output a five point summary table for multiple items
}
