% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_metrics_items_grouped}
\alias{tab_metrics_items_grouped}
\title{Output the means for groups in one or multiple columns}
\usage{
tab_metrics_items_grouped(
  data,
  cols,
  cols_groups,
  values = c("mean", "sd"),
  digits = 1,
  .negative = F,
  .labels = T
)
}
\arguments{
\item{data}{A tibble}

\item{cols}{The item columns that hold the values to summarize}

\item{values}{The output metrics, mean or sd}

\item{digits}{The digits to print}

\item{.negative}{If True (default), negative values are recoded to missing values}

\item{.labels}{If True (default) extracts item labels from the attributes, see get_labels()}

\item{cols_group}{The columns holding groups to compare}
}
\description{
TODO: handle completely missing data in single groups
}
