% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_metrics_one_cor}
\alias{tab_metrics_one_cor}
\title{Correlate two columns}
\usage{
tab_metrics_one_cor(
  data,
  cols,
  cross,
  negative = FALSE,
  method = "p",
  stats = FALSE,
  digits = 1,
  labels = TRUE,
  clean = TRUE,
  ...
)
}
\arguments{
\item{data}{A tibble}

\item{cross}{The column to correlate}

\item{negative}{If FALSE (default), negative values are recoded as missing values.}

\item{method}{The output metrics, p = Pearson's R, s = Spearman's rho}

\item{stats}{Add significance stars}

\item{digits}{The number of digits to print}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{tab_counts}.}

\item{col}{The first item}
}
\value{
A volker tibble
}
\description{
Correlate two columns
}
\keyword{internal}
