% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_metrics_one_grouped}
\alias{tab_metrics_one_grouped}
\title{Output a five point summary for groups}
\usage{
tab_metrics_one_grouped(
  data,
  col,
  col_group,
  negative = F,
  digits = 1,
  labels = T,
  clean = T,
  ...
)
}
\arguments{
\item{data}{A tibble}

\item{col}{The column holding metric values}

\item{col_group}{The column holding groups to compare}

\item{negative}{If FALSE (default), negative values are recoded as missing values.}

\item{digits}{The number of digits to print}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{tab_metrics}.}
}
\description{
Output a five point summary for groups
}
