% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_multi_corr}
\alias{tab_multi_corr}
\title{Output a correlation table}
\usage{
tab_multi_corr(data, cols1, cols2, method = "p", significant = F, digits = 2)
}
\arguments{
\item{data}{A tibble}

\item{cols1}{The source items}

\item{cols2}{The target items or NULL to calculate correlations within the source items}

\item{method}{The output metrics, p = Pearson's R, s = Spearman's rho}

\item{significant}{Only show significant values}

\item{digits}{The digits to print}
}
\description{
Output a correlation table
}
