% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_multi_means}
\alias{tab_multi_means}
\title{Output the means for groups in one or multiple columns}
\usage{
tab_multi_means(
  data,
  cols,
  cols_groups,
  values = c("mean", "sd"),
  digits = 1,
  .quiet = F
)
}
\arguments{
\item{data}{A tibble}

\item{cols}{The item columns that hold the values to summarize}

\item{values}{The output metrics, mean or sd}

\item{digits}{The digits to print}

\item{.quiet}{Set to true to suppress printing the output}

\item{cols_group}{The columns holding groups to compare}
}
\description{
TODO: handle completely missing data in single groups
}
