% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_var_metrics}
\alias{tab_var_metrics}
\alias{tab_metrics_var}
\title{Output a five point summary table for the values in multiple columns}
\usage{
tab_var_metrics(data, col, digits = 1, .labels = T)

tab_metrics_var(data, col, digits = 1, .labels = T)
}
\arguments{
\item{data}{A tibble}

\item{col}{The columns holding metric values}

\item{digits}{The digits to print}

\item{.labels}{If True (default) extracts item labels from the attributes, see get_labels()}
}
\description{
Output a five point summary table for the values in multiple columns

Alias for tab_var_metrics
}
