% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_compare.R
\name{table_compare_factor}
\alias{table_compare_factor}
\title{Compare factor value by group}
\usage{
table_compare_factor(data, col_category, col_group)
}
\arguments{
\item{data}{A tibble}

\item{col_category}{The column holding factor values}

\item{col_group}{The column holding groups to compare}
}
\description{
Compare factor value by group
}
