% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_compare.R
\name{table_compare_items}
\alias{table_compare_items}
\title{Table items}
\usage{
table_compare_items(data, cols_items, col_group)
}
\arguments{
\item{data}{A tibble containing item measures}

\item{cols_items}{Tidyselect item variables (e.g. starts_with...)}

\item{col_group}{Optional faceting variable}
}
\description{
Table items
}
