% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{theme_vlkr}
\alias{theme_vlkr}
\title{Define a default theme for volker plots}
\usage{
theme_vlkr(
  base_size = 11,
  base_color = "black",
  base_fill = list(VLKR_FILLCOLOR, VLKR_FILLDISCRETE)
)
}
\arguments{
\item{base_size}{Base font size}

\item{base_color}{Base font color}

\item{base_fill}{A list of fill color sets. Each set can contain different numbers of colors.}
}
\value{
A theme function
}
\description{
Define a default theme for volker plots
}
\examples{
library(volker)
library(ggplot2)
data <- volker::chatgpt

theme_set(theme_vlkr(base_size=15, base_fill = list("red"))
plot_counts(data, sd_gender)
}
