% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{tidy_lm_levels}
\alias{tidy_lm_levels}
\title{Tidy lm results, replace categorical parameter names by their levels and add the reference level}
\usage{
tidy_lm_levels(fit)
}
\arguments{
\item{fit}{Result of a \code{\link[stats:lm]{lm}} call.}
}
\value{
A tibble with regression parameters.
}
\description{
Tidy lm results, replace categorical parameter names by their levels and add the reference level
}
\author{
Created with the help of ChatGPT.
}
\keyword{internal}
