% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{trim_label}
\alias{trim_label}
\title{Remove trailing zeros and trailing or leading
whitespaces, colons,
hyphens and underscores}
\usage{
trim_label(x)
}
\arguments{
\item{x}{A character value}
}
\value{
The trimmed character value
}
\description{
Remove trailing zeros and trailing or leading
whitespaces, colons,
hyphens and underscores
}
\keyword{internal}
