% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{trunc_labels}
\alias{trunc_labels}
\title{Truncate labels}
\usage{
trunc_labels(x, max_length = 20)
}
\arguments{
\item{x}{A character vector.}

\item{max_length}{Specifies max_length, default is 20.}
}
\value{
A function that truncates labels to the max_length,
appending "..." if shortened.
}
\description{
Truncate labels that exceed a specified maximum length.
TODO: don't return function, use lambda expression in scale_x_discrete call
}
\keyword{internal}
