% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{vlkr_alpha_interpolated}
\alias{vlkr_alpha_interpolated}
\title{Interpolate an alpha value based on case numbers}
\usage{
vlkr_alpha_interpolated(
  n,
  n_min = 20,
  n_max = 100,
  alpha_min = VLKR_POINT_ALPHA,
  alpha_max = 1
)
}
\arguments{
\item{n}{Number of cases}

\item{n_min}{The case number where the minimum alpha value starts}

\item{n_max}{The case number where the maximum alpha value ends}

\item{alpha_min}{The minimum alpha value}

\item{alpha_max}{The maximum alpha value}
}
\value{
A value between the minimum and the maximum alpha value
}
\description{
Interpolate an alpha value based on case numbers
}
\keyword{internal}
