% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{vlkr_colors_contrast}
\alias{vlkr_colors_contrast}
\title{Given a vector of hex fill colors, choose an appropriate color for text}
\usage{
vlkr_colors_contrast(colors, threshold = 0.5)
}
\arguments{
\item{colors}{A vector of hex colors}

\item{threshold}{Luminance theshold for choosing black over white}
}
\value{
A vector of the same length as fill_colors with white or black colors.
}
\description{
Given a vector of hex fill colors, choose an appropriate color for text
}
\keyword{internal}
