% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{vlkr_colors_discrete}
\alias{vlkr_colors_discrete}
\title{Get colors for discrete scales}
\usage{
vlkr_colors_discrete(n)
}
\arguments{
\item{n}{Number of colors.}
}
\value{
A vector of colors.
}
\description{
If the option ggplot2.discrete.fill is set,
gets color values from the first list item that
has enough colors and reverses them to start filling
from the left in grouped bar charts.
}
\details{
Falls back to scale_fill_hue().
}
\keyword{internal}
