% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{vlkr_colors_sequential}
\alias{vlkr_colors_sequential}
\title{Get colors for sequential scales}
\usage{
vlkr_colors_sequential(n)
}
\arguments{
\item{n}{Number of colors.}
}
\value{
A vector of colors.
}
\description{
Creates a gradient scale based on VLKR_FILLGRADIENT.
}
\keyword{internal}
