% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{VLKR_FIG_SETTINGS}
\alias{VLKR_FIG_SETTINGS}
\title{Resolution settings for plots}
\format{
An object of class \code{list} of length 2.
}
\usage{
VLKR_FIG_SETTINGS
}
\description{
Override with \code{options(vlkr.fig.settings=list(html = list(dpi = 192, scale = 2, width = 910, pxperline = 15)))}.
Add a key for each output format when knitting a document.
You can override the width by setting vlkr.fig.width in the chunk options.
}
\keyword{internal}
