% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{VLKR_FILLDISCRETE}
\alias{VLKR_FILLDISCRETE}
\title{Fill colors}
\format{
An object of class \code{list} of length 3.
}
\usage{
VLKR_FILLDISCRETE
}
\description{
Override with \code{options(vlkr.discrete.fill=list(c("purple")))}.
}
\keyword{internal}
