% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{VLKR_FILLGRADIENT}
\alias{VLKR_FILLGRADIENT}
\title{Gradient colors}
\format{
An object of class \code{character} of length 5.
}
\usage{
VLKR_FILLGRADIENT
}
\description{
Gradient colors
}
\keyword{internal}
