% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{VLKR_FILLPOLARIZED}
\alias{VLKR_FILLPOLARIZED}
\title{Polarized colors}
\format{
An object of class \code{character} of length 5.
}
\usage{
VLKR_FILLPOLARIZED
}
\description{
Polarized colors
}
\keyword{internal}
