% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{VLKR_MAX_CATEGORIES}
\alias{VLKR_MAX_CATEGORIES}
\title{Maximum number of distinct values to determine whether a column selection
contains only categorical values}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
VLKR_MAX_CATEGORIES
}
\description{
Override with \code{options(vlkr.max.categories=10)}.
}
\keyword{internal}
