% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{VLKR_NA_LEVELS}
\alias{VLKR_NA_LEVELS}
\title{Levels to remove from factors}
\format{
An object of class \code{character} of length 1.
}
\usage{
VLKR_NA_LEVELS
}
\description{
By default the value \verb{[NA] nicht beantwortet} is removed from tables and plots
}
\keyword{internal}
