% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{VLKR_NA_NUMBERS}
\alias{VLKR_NA_NUMBERS}
\title{Numbers to remove from vectors}
\format{
An object of class \code{numeric} of length 3.
}
\usage{
VLKR_NA_NUMBERS
}
\description{
Override with \code{options(vlkr.na.numbers=c(-2,-9))}.
}
\keyword{internal}
