% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{VLKR_NA_OMIT}
\alias{VLKR_NA_OMIT}
\title{Whether to remove all missings}
\format{
An object of class \code{logical} of length 1.
}
\usage{
VLKR_NA_OMIT
}
\description{
Override with \code{options(vlkr.na.omit=FALSE)} to use all cases if possible.
}
\keyword{internal}
