% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{VLKR_NORMAL_DIGITS}
\alias{VLKR_NORMAL_DIGITS}
\title{Output thresholds}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
VLKR_NORMAL_DIGITS
}
\description{
Output thresholds
}
\keyword{internal}
