% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{VLKR_PLOT_RESOLUTION}
\alias{VLKR_PLOT_RESOLUTION}
\title{Resolution settings for plots}
\format{
An object of class \code{list} of length 2.
}
\usage{
VLKR_PLOT_RESOLUTION
}
\description{
Add a key for each output format when knitting a document.
}
\keyword{internal}
