% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{VLKR_POINT_ALPHA}
\alias{VLKR_POINT_ALPHA}
\title{Alpha values}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
VLKR_POINT_ALPHA
}
\description{
Alpha values
}
\keyword{internal}
