% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{VLKR_POINT_MEAN_SHAPE}
\alias{VLKR_POINT_MEAN_SHAPE}
\title{Shapes}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
VLKR_POINT_MEAN_SHAPE
}
\description{
Shapes
}
\keyword{internal}
