% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{VLKR_WRAP_SEPARATOR}
\alias{VLKR_WRAP_SEPARATOR}
\title{Word wrap separators}
\format{
An object of class \code{character} of length 1.
}
\usage{
VLKR_WRAP_SEPARATOR
}
\description{
Word wrap separators
}
\keyword{internal}
