% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volker-package.R
\docType{package}
\name{volker-package}
\alias{volker}
\alias{volker-package}
\title{volker: Ein Package für Übersichten und Grafiken zu Befragungsdaten}
\description{
Dieses Package enthält Funktionen zur Erstellung von Übersichten und Grafiken zu Befragungsdaten.
}
\author{
\strong{Maintainer}: Datavana Team \email{jakob.juenger@uni-muenster.de}

}
\keyword{internal}
