% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{wrap_label}
\alias{wrap_label}
\title{Wrap a string}
\usage{
wrap_label(x, width = 40)
}
\arguments{
\item{x}{A character vector.}

\item{width}{The number of chars after which to break.}
}
\value{
A character vector with wrapped strings.
}
\description{
Wrap a string
}
\keyword{internal}
