% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zip_tables}
\alias{zip_tables}
\title{Combine two identically shaped data frames
by adding values of each column from the second data frame
into the corresponding column in the first dataframe using parentheses}
\usage{
zip_tables(x, y, newline = TRUE, brackets = FALSE)
}
\arguments{
\item{x}{The first data frame}

\item{y}{The second data frame}

\item{newline}{Whether to add a new line character between the values (default: TRUE).}

\item{brackets}{Whether to set the secondary values in brackets (default: FALSE).}
}
\value{
A combined data frame
}
\description{
TODO: implement newline parameter
}
\examples{
library(tibble)

tab_n <- tibble(item=c("Female", "Male"), n=c(15,10))
tab_p <- tibble( item=c("Female", "Male"), n=c("60\%","30\%"))

volker:::zip_tables(tab_n, tab_p)
}
\keyword{internal}
