% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx.R
\name{add_idx}
\alias{add_idx}
\title{Calculate the mean value of multiple items}
\usage{
add_idx(data, newcol, cols, .negative = FALSE)
}
\arguments{
\item{data}{A dataframe}

\item{newcol}{Name of the index as a character value}

\item{cols}{A tidy selection of item columns}

\item{.negative}{If TRUE, all negative values are recoded to NA}
}
\description{
Calculate the mean value of multiple items
}
