% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{cfg_get_na_numbers}
\alias{cfg_get_na_numbers}
\title{Get configured na numbers}
\usage{
cfg_get_na_numbers()
}
\description{
Retrieves values either from the option or from the constant.
}
\keyword{A}
\keyword{NAs}
\keyword{as}
\keyword{be}
\keyword{internal}
\keyword{numbers}
\keyword{return}
\keyword{should}
\keyword{that}
\keyword{treated}
\keyword{vector}
\keyword{with}
