% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_has_column}
\alias{check_has_column}
\title{Check whether a column exist and stop if not}
\usage{
check_has_column(data, cols, msg = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{A tidyselection of columns.}

\item{msg}{A custom error message if the check fails.}
}
\value{
boolean Whether the column exists.
}
\description{
Check whether a column exist and stop if not
}
\keyword{internal}
