% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_columns}
\alias{clean_columns}
\title{Prepare metadata and types of dataframe columns}
\usage{
clean_columns(data, remove.na = T)
}
\arguments{
\item{data}{Data frame}

\item{remove.na}{Whether residual values should be replaced by NA}
}
\value{
Data frame
}
\description{
Prepares Sosci data:
\itemize{
\item Remove the avector class from all columns
(comes from Sosci and prevents combining vectors)
\item Recode residual factor values to NA ("\link{NA} nicht beantwortet")
\item Recode -9 values to NA
}
}
