% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_plot}
\alias{cluster_plot}
\title{Get plot for clustering result}
\usage{
cluster_plot(
  data,
  cols,
  newcol = NULL,
  k = NULL,
  method = NULL,
  type = "lines",
  reorder = TRUE,
  labels = TRUE,
  clean = TRUE,
  ...
)
}
\arguments{
\item{data}{A tibble.}

\item{cols}{A tidy selection of item columns or a single column with cluster values as a factor or logical.
If the column already contains a cluster result from \link{add_clusters}, it is used, and other parameters are ignored.
If no cluster result exists, it is calculated with \link{add_clusters}.}

\item{newcol}{Name of the new cluster column as a character vector.
Set to NULL (default) to automatically build a name
from the common column prefix, prefixed with "cls_".}

\item{k}{Number of clusters to calculate.
Set to NULL to automatically determine an optimal cluster number.
For kmeans, outputs a scree plot based on within-sums of squares for up to 10 clusters.
In this case, the number of clusters is automatically chosed based on the elbow criterion.
For pam, outputs a silhoette plot for up to 10 clusters.
In this case, the number of clusters is automatically chosen based on the maximum average silhouette.}

\item{method}{The method as a character value, one of \code{kmeans} or \code{pam}. See \link{add_clusters} for further details.}

\item{type}{The plot type, one of \code{"lines"} or \code{"heatmap"}.}

\item{reorder}{Reorder items to minimize line crossings,
Either \code{TRUE} to automatically select a
method (\code{"olo"} if \pkg{seriation} is installed, otherwise \code{"min"}), or
one of the character values \code{"max"}, \code{"min"}, \code{"spread"}, \code{"gw"}, or
\code{"olo"}. Defaults to \code{FALSE} which disables reordering.}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{plot_metrics}.}
}
\value{
A ggplot object.
}
\description{
Clustering is performed using \link{add_clusters}.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
library(volker)
data <- volker::chatgpt

cluster_plot(data, starts_with("cg_adoption"), k = 2)

}
\keyword{internal}
