% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{codebook}
\alias{codebook}
\title{Get variable labels from their comment attributes}
\usage{
codebook(data, cols)
}
\arguments{
\item{data}{A tibble.}

\item{cols}{A tidy variable selections to filter specific columns.}
}
\value{
A tibble with the columns:
- item_name: The column name.
- item_group: First part of the column name, up to an underscore.
- item_class: The last class value of an item (e.g. numeric, factor).
- item_label: The comment attribute of the column.
- value_name: In case a column has numeric attributes, the attribute names.
- value_label: In case a column has numeric attributes or T/F-attributes,
the attribute values.
In case a column has a levels attribute, the levels.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
volker::codebook(volker::chatgpt)
}
