% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{data_cat}
\alias{data_cat}
\title{Convert numeric values to string}
\usage{
data_cat(data, cols)
}
\arguments{
\item{data}{A data frame containing the items to be converted.}

\item{cols}{A tidy selection of columns to convert.}
}
\value{
A data frame with the converted values
}
\description{
Convert numeric values to string
}
\keyword{internal}
