% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{data_clean}
\alias{data_clean}
\title{Prepare dataframe for tabs, plots, and index operations}
\usage{
data_clean(data, plan = "sosci", ...)
}
\arguments{
\item{data}{Data frame}

\item{plan}{The cleaning plan. By now, only "sosci" is supported. See \link{data_clean_sosci}.}

\item{...}{Other parameters passed to the appropriate cleaning function}
}
\value{
Cleaned data frame with vlkr_df class
}
\description{
The tibble remembers whether it was already cleaned and
the cleaning plan is only performed once in the first call.
}
\examples{
ds <- volker::chatgpt
ds <- data_clean(ds)
}
