% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{data_prepare}
\alias{data_prepare}
\title{Prepare data for calculation}
\usage{
data_prepare(
  data,
  cols,
  cross,
  negative = TRUE,
  clean = TRUE,
  rm_neg_cols = FALSE
)
}
\arguments{
\item{data}{Data frame to be prepared.}

\item{cols}{The first column selection.}

\item{cross}{The second column selection.}

\item{negative}{Whether to remove negatives.}

\item{clean}{Whether to clean data using \link{data_clean}.}

\item{rm_neg_cols}{if TRUE and cross is specified, negative values will be removed only from the cols.}
}
\value{
Prepared data frame.
}
\description{
Clean data, check column selection, remove missing and negatives.
}
\examples{
data <- volker::chatgpt
data_prepare(data, sd_age, sd_gender)

}
\keyword{internal}
