% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{data_prepare}
\alias{data_prepare}
\title{Prepare data for calculation}
\usage{
data_prepare(
  data,
  cols,
  cross,
  rm.missings = TRUE,
  rm.negatives = FALSE,
  clean = TRUE
)
}
\arguments{
\item{data}{Data frame to be prepared.}

\item{cols}{The first column selection.}

\item{cross}{The second column selection.}

\item{rm.missings}{Whether to remove cases with missing values (default TRUE).}

\item{rm.negatives}{Whether to remove negatives in all columns (TRUE),
no columns (FALSE), or only the columns in the cols parameter ("cols").}

\item{clean}{Whether to clean data using \link{data_clean}.}
}
\value{
Prepared data frame.
}
\description{
Clean data, check column selection, remove missing and negatives.
}
\examples{
data <- volker::chatgpt
data_prepare(data, sd_age, sd_gender)

}
\keyword{internal}
