% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree.R
\name{.agree_items}
\alias{.agree_items}
\title{Calculate agreement coefficients for multiple items}
\usage{
.agree_items(
  data,
  cols,
  coders,
  ids = NULL,
  category = NULL,
  method = "reliability",
  labels = TRUE
)
}
\arguments{
\item{data}{A tibble.}

\item{cols}{The columns holding codings.}

\item{coders}{The column holding coders.}

\item{ids}{The column holding case identifiers.}

\item{method}{The output metrics, one of reliability (reliability scores)
or classification (accuracy, precision, recall, f1).}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}
}
\value{
A tibble with agreement coefficients.
}
\description{
Calculate agreement coefficients for multiple items
}
\keyword{internal}
