% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.attr_transfer}
\alias{.attr_transfer}
\title{Transfer an attribute from one to another object}
\usage{
.attr_transfer(to, from, key)
}
\arguments{
\item{to}{The target object.}

\item{from}{The source object.}

\item{key}{The attribute key as character value.}
}
\value{
The target object with the updated attribute.
}
\description{
Transfer an attribute from one to another object
}
\keyword{internal}
