% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{.effect_npmi}
\alias{.effect_npmi}
\title{Calculate nmpi}
\usage{
.effect_npmi(data, col, cross, labels = TRUE, clean = TRUE, smoothing = 0, ...)
}
\arguments{
\item{data}{A tibble.}

\item{col}{The column holding factor values.}

\item{cross}{The column to correlate.}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{smoothing}{Add pseudocount. Calculate the pseudocount based on the number of trials
to apply Laplace's rule of succession.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{tab_counts}.}
}
\value{
A volker tibble.
}
\description{
Calculate nmpi
}
\keyword{internal}
