% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_bars}
\alias{.plot_bars}
\title{Helper function: plot grouped bar chart}
\usage{
.plot_bars(
  data,
  category = NULL,
  scale = NULL,
  numbers = NULL,
  title = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{Dataframe with the columns Item, value, p, n}

\item{category}{Category for filtering the dataframe}

\item{scale}{Direction of the scale: 0 = no direction for categories,
-1 = descending or 1 = ascending values.}

\item{numbers}{The values to print on the bars: "n" (frequency), "p" (percentage) or both.}

\item{title}{The plot title or NULL}

\item{caption}{The plot caption or NULL. The caption is used for notes.}
}
\description{
Helper function: plot grouped bar chart
}
