% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_heatmap}
\alias{.plot_heatmap}
\title{Helper function: Heatmap}
\usage{
.plot_heatmap(result, value_col, base_n, numbers, title, labels)
}
\arguments{
\item{result}{A tibble with item combinations in the first two columns and a correlation value in the last column}

\item{value_col}{The name of the column containing correlation values, a character value}

\item{numbers}{Controls whether to display correlation coefficients on the plot.}

\item{title}{If TRUE (default) shows a plot title derived from the column labels.
Disable the title with FALSE or provide a custom title as character value.}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}
}
\value{
A ggplot object
}
\description{
Helper function: Heatmap
}
\keyword{internal}
