% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_lines}
\alias{.plot_lines}
\title{Helper function: plot grouped line chart}
\usage{
.plot_lines(data, scale = NULL, base = NULL, title = NULL)
}
\arguments{
\item{data}{Dataframe with the columns item, value}

\item{scale}{???}

\item{base}{The plot base as character or NULL.}

\item{title}{The plot title as character or NULL}
}
\value{
A ggplot object
}
\description{
Helper function: plot grouped line chart
}
\keyword{internal}
