% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{.report_mdl}
\alias{.report_mdl}
\title{Generate an model table and plot}
\usage{
.report_mdl(data, cols, categorical = NULL, metric = NULL, ..., title = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{A a single column (without quotes).}

\item{categorical}{A tidy column selection holding independet categorical variables.}

\item{metric}{A tidy column selection holding independent metric variables.}

\item{title}{Add a plot title (default = TRUE).}
}
\value{
A list containing a table and a plot volker report chunk.
}
\description{
Generate an model table and plot
}
\keyword{internal}
