% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.to_vlkr_plot}
\alias{.to_vlkr_plot}
\title{Add the volker class and options}
\usage{
.to_vlkr_plot(
  pl,
  rows = NULL,
  maxlab = NULL,
  baseline = TRUE,
  theme_options = TRUE
)
}
\arguments{
\item{pl}{A ggplot object.}

\item{rows}{The number of items on the vertical axis. Will be automatically determined when NULL.
For stacked bar charts, don't forget to set the group parameter, otherwise it won't work}

\item{maxlab}{The character length of the longest label to be plotted. Will be automatically determined when NULL.
on the vertical axis.}

\item{baseline}{Whether to print a message about removed values.}

\item{theme_options}{Enable or disable axis titles and text, by providing a list with any of the elements
axis.text.x, axis.text.y, axis.title.x, axis.title.y set to TRUE or FALSE.
By default, titles (=scale labels) are disabled and text (= the tick labels) are enabled.}
}
\value{
A ggplot object with vlkr_plt class.
}
\description{
Add the volker class and options
}
\keyword{internal}
