% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{.to_vlkr_tab}
\alias{.to_vlkr_tab}
\title{Add vlkr_tbl class}
\usage{
.to_vlkr_tab(data, digits = NULL)
}
\arguments{
\item{data}{A tibble}

\item{digits}{Set the plot digits. If NULL (default), no digits are set.}
}
\value{
A volker tibble
}
\description{
Additionally, removes the skim_df class if present
}
\keyword{internal}
