% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{effect_counts}
\alias{effect_counts}
\title{Output effect sizes for count data}
\usage{
effect_counts(data, cols, cross = NULL, metric = FALSE, clean = TRUE, ...)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{A tidy column selection,
e.g. a single column (without quotes)
or multiple columns selected by methods such as starts_with().}

\item{cross}{Optional, a grouping column. The column name without quotes.}

\item{metric}{When crossing variables, the cross column parameter can contain categorical or metric values.
By default, the cross column selection is treated as categorical data.
Set metric to TRUE, to treat it as metric and calculate correlations.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Other parameters passed to the appropriate stat function.}
}
\value{
A volker tibble.
}
\description{
The type of effect size depends on the number of selected columns:
\itemize{
\item One column: see \link{effect_counts_one}
\item Multiple columns: see \link{effect_counts_items}
\item One column and one grouping column: see \link{effect_counts_one_grouped}
\item Multiple columns and one grouping column: see \link{effect_counts_items_grouped}
}
}
\examples{
library(volker)
data <- volker::chatgpt

effect_counts(data, sd_gender, adopter)

}
