% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{effect_counts_items_grouped}
\alias{effect_counts_items_grouped}
\title{Effect size and test for comparing multiple variables by a grouping variable}
\usage{
effect_counts_items_grouped(data, cols, cross, clean = TRUE, ...)
}
\arguments{
\item{data}{A tibble containing item measures and grouping variable.}

\item{cols}{Tidyselect item variables (e.g. starts_with...).}

\item{cross}{The column holding groups to compare.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{effect_counts}.}
}
\value{
A volker tibble.
}
\description{
\strong{Not yet implemented. The future will come.}
}
\keyword{internal}
