% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{effect_counts_one}
\alias{effect_counts_one}
\title{Test whether the shares differ from homogeneity}
\usage{
effect_counts_one(
  data,
  col,
  digits = 2,
  percent = TRUE,
  labels = TRUE,
  clean = TRUE,
  ...
)
}
\arguments{
\item{data}{A tibble.}

\item{col}{The column holding factor values.}

\item{digits}{The number of digits to print.}

\item{percent}{Proportions are formatted as percent by default. Set to FALSE to get bare proportions.}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{effect_counts}.}
}
\value{
A volker tibble.
}
\description{
\strong{Not yet implemented. The future will come.}
}
\keyword{internal}
