% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{effect_counts_one}
\alias{effect_counts_one}
\title{Test homogeneity of category shares}
\usage{
effect_counts_one(data, col, clean = TRUE, ...)
}
\arguments{
\item{data}{A tibble.}

\item{col}{The column holding factor values.}

\item{clean}{Prepare data by \link{data_clean}}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{effect_counts}.}
}
\value{
A volker tibble.
}
\description{
Performs a goodness-of-fit test and calculates the Gini coefficient.
The goodness-of-fit-test is calculated using \code{stats::\link[stats:chisq.test]{chisq.test}}.
}
\examples{
library(volker)
data <- volker::chatgpt
data <- data |>
  filter(sd_gender != "diverse") |>
  effect_counts_one(sd_gender)

}
\keyword{internal}
