% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{effect_metrics_items}
\alias{effect_metrics_items}
\title{Output test statistics and effect size (Cohen's d) for paired samples}
\usage{
effect_metrics_items(
  data,
  cols,
  method = "pearson",
  negative = FALSE,
  labels = TRUE,
  clean = TRUE,
  ...
)
}
\arguments{
\item{data}{A tibble containing item measures.}

\item{cols}{The column holding metric values.}

\item{method}{The output metrics, TRUE or pearson = Pearson's R, spearman = Spearman's rho.}

\item{negative}{If FALSE (default), negative values are recoded as missing values.}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{effect_metrics}.}
}
\value{
A volker table containing correlations.
}
\description{
Output test statistics and effect size (Cohen's d) for paired samples
}
\examples{
library(volker)
data <- volker::chatgpt

effect_metrics_items(data, starts_with("cg_adoption"))


}
\keyword{internal}
