% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{get_baseline}
\alias{get_baseline}
\title{Get a formatted baseline for removed zero, negative, and missing cases}
\usage{
get_baseline(obj)
}
\arguments{
\item{obj}{An object with the missings attribute.}
}
\value{
A formatted message or NULL if the missings attribute is not present.
}
\description{
Get a formatted baseline for removed zero, negative, and missing cases
}
\keyword{internal}
