% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{get_baseline}
\alias{get_baseline}
\title{Get a formatted baseline from attributes of an object.}
\usage{
get_baseline(obj, ignore = c())
}
\arguments{
\item{obj}{An object with supported attributes.}

\item{ignore}{Characer vector of attributes to ignore.}
}
\value{
A formatted message or NULL if none of the supported attributes is present.
}
\description{
The following attributes are considered:
\itemize{
\item cases: Number of cases.
\item missing: Removed zero, negative, and missing cases.
\item focus: Focus category.
\item auto: The k value of cluster methods.
\item reversed: A list of reversed items.
\item split: Items that were split at the median.
\item adjust: The adjustment method for p values.
}
}
\keyword{internal}
