% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{get_baseline}
\alias{get_baseline}
\title{Get a formatted baseline for removed zero, negative, and missing cases
and include focus category information if present}
\usage{
get_baseline(obj)
}
\arguments{
\item{obj}{An object with the missings and focus attributes.}
}
\value{
A formatted message or NULL if missings and focus attributes are not present.
}
\description{
Get a formatted baseline for removed zero, negative, and missing cases
and include focus category information if present
}
\keyword{internal}
