% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{get_labels}
\alias{get_labels}
\title{Get the labels of values from a codebook}
\usage{
get_labels(codes, values)
}
\arguments{
\item{codes}{The codebook as it results from the codebook() function}

\item{values}{A vector of labels}
}
\value{
The labels. If the values are not present in the codebook, returns the values.
}
\description{
Get the labels of values from a codebook
}
\keyword{internal}
