% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_labels}
\alias{get_labels}
\title{Get variable labels from their comment attributes}
\usage{
get_labels(data, cols)
}
\arguments{
\item{data}{A tibble}
}
\value{
A tibble with the columns item_name, item_label, value_name, and value_label
}
\description{
Get variable labels from their comment attributes
}
