% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_labels}
\alias{get_labels}
\title{Get variable labels from their comment attributes}
\usage{
get_labels(data)
}
\arguments{
\item{data}{A tibble}
}
\description{
Get variable labels from their comment attributes
}
