% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{get_limits}
\alias{get_limits}
\title{Get the numeric range from the labels}
\usage{
get_limits(data, cols, negative = TRUE)
}
\arguments{
\item{data}{The labeled data frame.}

\item{cols}{A tidy variable selection.}

\item{negative}{Whether to include negative values.}
}
\value{
A list or NULL.
}
\description{
Gets the range of all values in the selected columns
by the first successful of the following methods:
}
\details{
\itemize{
\item Inspect the limits column attribute.
\item Lookup the value names in the codebook.
\item Calculate the range from all values in the columns.
}
}
\keyword{internal}
