% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{get_prefix}
\alias{get_prefix}
\title{Get the common prefix of character values}
\usage{
get_prefix(x, ignore.case = FALSE, trim = FALSE)
}
\arguments{
\item{x}{Character vector}

\item{ignore.case}{Whether case matters (default)}

\item{trim}{Whether non alphabetic characters should be trimmed}
}
\value{
The longest common prefix of the strings
}
\description{
Helper function taken from the biobase package.
Duplicated here instead of loading the package to avoid overhead.
See https://github.com/Bioconductor/Biobase
}
\keyword{internal}
