% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx.R
\name{idx_add}
\alias{idx_add}
\title{Deprecated Alias for \code{add_index}}
\usage{
idx_add(data, cols, newcol = NULL, reverse = NULL, clean = TRUE)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{idx_add()} was renamed to \code{add_index()}.
}
\details{
This function is a deprecated alias for \code{add_index}.
}
\keyword{internal}
