% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{knit_table}
\alias{knit_table}
\title{Knit volker tables}
\usage{
knit_table(df, ...)
}
\arguments{
\item{df}{Data frame}
}
\value{
Formatted  table produced by \link{kable}
}
\description{
Knit volker tables
}
\keyword{internal}
