% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labs_apply}
\alias{labs_apply}
\title{Set variable labels by setting their comment attributes}
\usage{
labs_apply(data, codes, values = T)
}
\arguments{
\item{data}{A tibble}

\item{codes}{A tibble in \link{codebook} format.
To set column labels, use item_name and item_label columns.}

\item{values}{If TRUE (default), sets value labels.
- For item values:
Value labels are retrieved from the columns
value_name and value_label in your codebook.
- For factors: Factor levels and order are retrieved
from the value_label column.}
}
\value{
A tibble with new labels
}
\description{
Set variable labels by setting their comment attributes
}
