% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labs_store}
\alias{labs_store}
\title{Get the current codebook and store it in the codebook attribute.}
\usage{
labs_store(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame
}
\description{
You can restore the labels after mutate operations by calling
\link{labs_restore}.
}
\examples{
library(dplyr)
library(volker)

volker::chatgpt |>
  labs_store() |>
  mutate(sd_age = 2024 - sd_age) |>
  labs_restore() |>
  tab_metrics(sd_age)
}
