% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_compare_factor}
\alias{plot_compare_factor}
\title{Compare and plot factor value by group}
\usage{
plot_compare_factor(data, col_category, col_group, relative = F)
}
\arguments{
\item{data}{A tibble containing data}

\item{col_category}{The column holding factor values}

\item{col_group}{The column holding groups to compare}

\item{relative}{Show absolute (F) or relative values (T)}
}
\description{
Compare and plot factor value by group
}
