% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_counts}
\alias{plot_counts}
\title{Compare and plot items}
\usage{
plot_counts(data, cols_items, col_group)
}
\arguments{
\item{data}{A tibble containing item measures}

\item{cols_items}{Tidyselect item variables (e.g. starts_with...)}

\item{col_group}{Optional faceting variable}
}
\description{
Compare and plot items
}
