% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_counts_items_grouped}
\alias{plot_counts_items_grouped}
\title{Plot frequencies of multiple items compared by groups}
\usage{
plot_counts_items_grouped(data, cols, col_group, clean = TRUE, ...)
}
\arguments{
\item{data}{A tibble}

\item{cols}{The item columns that hold the values to summarize}

\item{col_group}{The column holding groups to compare}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{plot_counts}.}
}
\value{
A ggplot object
}
\description{
TODO: implement -> focus one category and show n / p
}
\keyword{internal}
