% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_counts_one_cor}
\alias{plot_counts_one_cor}
\title{Plot correlation of categories with one metric column}
\usage{
plot_counts_one_cor(
  data,
  col,
  cross,
  ordered = NULL,
  limits = NULL,
  missings = FALSE,
  numbers = NULL,
  title = TRUE,
  labels = TRUE,
  clean = TRUE,
  ...
)
}
\arguments{
\item{data}{A tibble.}

\item{col}{The column holding factor values.}

\item{cross}{The metric column}

\item{ordered}{Values can be nominal (0) or ordered ascending (1) descending (-1).
By default (NULL), the ordering is automatically detected.
An appropriate color scale should be choosen depending on the ordering.
For unordered values, colors from VLKR_FILLDISCRETE are used.
For ordered values, shades of the VLKR_FILLGRADIENT option are used.}

\item{limits}{The scale limits, autoscaled by default.
Set to \code{c(0,100)} to make a 100 \% plot.}

\item{missings}{Include missing values (default FALSE)}

\item{numbers}{The numbers to print on the bars: "n" (frequency), "p" (percentage) or both.}

\item{title}{If TRUE (default) shows a plot title derived from the column labels.
Disable the title with FALSE or provide a custom title as character value.}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{plot_counts}.}
}
\value{
A ggplot object
}
\description{
\strong{Not yet implemented. The future will come.}
}
\keyword{internal}
