% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_counts_one_grouped}
\alias{plot_counts_one_grouped}
\title{Plot frequencies cross tabulated with a grouping column}
\usage{
plot_counts_one_grouped(
  data,
  col,
  col_group,
  numbers = NULL,
  prop = "total",
  ordered = NULL,
  missings = F,
  title = T,
  .labels = T,
  .category = NULL
)
}
\arguments{
\item{data}{A tibble}

\item{col}{The column holding factor values}

\item{col_group}{The column holding groups to compare}

\item{numbers}{The numbers to print on the bars: "n" (frequency), "p" (percentage) or both.}

\item{prop}{The basis of percent calculation: total (the default) or rows.
To display column proportions, swap the first column with the grouping column.}

\item{ordered}{Values can be nominal (0) or ordered ascending (1) descending (-1).
By default (NULL), the ordering is automatically detected.
An appropriate color scale should be choosen depending on the ordering.
For unordered values, the default scale is used.
For ordered values, shades of the VLKR_FILLGRADIENT option are used.}

\item{missings}{Include missing values (default FALSE)}

\item{.labels}{If TRUE (default) extracts item labels from the attributes, see get_labels()}

\item{.category}{Set a character value to focus only selected categories. In case of boolean values, automatically, only one category is plotted. Set to FALSE to plot all categories.}
}
\description{
Note: only non-missing cases are used to calculate the percentage.
}
