% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_group_counts}
\alias{plot_group_counts}
\title{Output frequencies for multiple variables}
\usage{
plot_group_counts(
  data,
  col,
  col_group,
  values = c("n", "p"),
  .labels = T,
  .binary = NULL
)
}
\arguments{
\item{data}{A tibble}

\item{col}{The column holding factor values}

\item{col_group}{The column holding groups to compare}

\item{values}{The values to print on the bars: n (frequency) or p (percentage).Not implemented yet.}

\item{.labels}{If True (default) extracts item labels from the attributes, see get_labels()}

\item{.binary}{Set a character value to focus only selected categories. In case of boolean values, automatically, only one category is plotted. Set to FALSE to plot all categories.}
}
\description{
Output frequencies for multiple variables
}
