% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_group_counts}
\alias{plot_group_counts}
\title{Plot frequencies cross tabulated with a grouping column}
\usage{
plot_group_counts(
  data,
  col,
  col_group,
  numbers = NULL,
  prop = "total",
  .labels = T,
  .category = NULL
)
}
\arguments{
\item{data}{A tibble}

\item{col}{The column holding factor values}

\item{col_group}{The column holding groups to compare}

\item{numbers}{The numbers to print on the bars: "n" (frequency), "p" (percentage) or both.}

\item{prop}{The basis of percent calculation: total (the default) or rows. To display column proportions, swap the first column and the grouping column.}

\item{.labels}{If True (default) extracts item labels from the attributes, see get_labels()}

\item{.category}{Set a character value to focus only selected categories. In case of boolean values, automatically, only one category is plotted. Set to FALSE to plot all categories.}
}
\description{
Note: only non-missing cases are used to calculate the percentage.
}
