% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_group_metrics}
\alias{plot_group_metrics}
\title{Output averages for multiple variables}
\usage{
plot_group_metrics(
  data,
  col,
  col_group,
  limits = NULL,
  numbers = NULL,
  .labels = T,
  .negative = F
)
}
\arguments{
\item{data}{A tibble containing item measures}

\item{col}{The column holding metric values}

\item{col_group}{The column holding groups to compare}

\item{limits}{The scale limits. Set NULL to extract limits from the labels.}

\item{numbers}{The values to print on the bars: "m" or NULL}

\item{.labels}{If True (default) extracts item labels from the attributes, see get_labels()}

\item{.negative}{If False (default) negative values are recoded as missing values}
}
\description{
Output averages for multiple variables
}
