% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_grouped_bars}
\alias{plot_grouped_bars}
\title{Helper function: plot grouped bar chart}
\usage{
plot_grouped_bars(data, .category = NULL, .direction = -1)
}
\arguments{
\item{data}{Dataframe with the columns Item, value, p, n}

\item{.category}{Category for filtering the dataframe}

\item{.direction}{Direction of the viridis scale, either -1 or 1.}
}
\description{
Helper function: plot grouped bar chart
}
