% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plot_item_counts}
\alias{plot_item_counts}
\title{Alias for plot_counts_items}
\usage{
plot_item_counts(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{plot_item_counts} has been deprecated. Use \link{plot_counts_items} instead.
}
\keyword{internal}
