% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_item_counts}
\alias{plot_item_counts}
\title{Output frequencies for multiple variables}
\usage{
plot_item_counts(data, cols, .numbers = NULL, .labels = T, .category = NULL)
}
\arguments{
\item{data}{A tibble containing item measures}

\item{cols}{Tidyselect item variables (e.g. starts_with...)}

\item{.numbers}{The values to print on the bars: "n" (frequency), "p" (percentage) or both.}

\item{.labels}{If True (default) extracts item labels from the attributes, see get_labels()}

\item{.category}{Set a character value to focus only selected categories. In case of boolean values, automatically, only one category is plotted. Set to FALSE to plot all categories.}
}
\description{
Note: only non-missing cases are used to calculate the percentage.
}
