% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_metrics_items_grouped}
\alias{plot_metrics_items_grouped}
\title{Output averages for multiple variables compared by a grouping variable}
\usage{
plot_metrics_items_grouped(
  data,
  cols,
  cols_groups,
  limits = NULL,
  numbers = NULL,
  title = T,
  .labels = T,
  .negative = F
)
}
\arguments{
\item{data}{A tibble containing item measures}

\item{cols}{Tidyselect item variables (e.g. starts_with...)}

\item{limits}{The scale limits. Set NULL to extract limits from the labels.}

\item{numbers}{The values to print on the bars: "m" or NULL}

\item{.labels}{If True (default) extracts item labels from the attributes, see get_labels()}

\item{.negative}{If False (default) negative values are recoded as missing values}

\item{cols_group}{The columns holding groups to compare}
}
\value{
A plot
}
\description{
Output averages for multiple variables compared by a grouping variable
}
