% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_metrics_items_grouped}
\alias{plot_metrics_items_grouped}
\title{Output averages for multiple variables compared by a grouping variable}
\usage{
plot_metrics_items_grouped(
  data,
  cols,
  col_group,
  limits = NULL,
  negative = F,
  title = T,
  labels = T,
  ...
)
}
\arguments{
\item{data}{A tibble containing item measures}

\item{cols}{Tidyselect item variables (e.g. starts_with...)}

\item{limits}{The scale limits. Set NULL to extract limits from the labels. NOT IMPLEMENTED YET.}

\item{negative}{If FALSE (default), negative values are recoded as missing values.}

\item{title}{If TRUE (default) shows a plot title derived from the column labels.
Disable the title with FALSE or provide a custom title as character value.}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}

\item{cols_group}{The column holding groups to compare}
}
\value{
A plot
}
\description{
Output averages for multiple variables compared by a grouping variable
}
