% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{prepare}
\alias{prepare}
\title{Prepare data}
\usage{
prepare(data, ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

The function \code{prepare} has been deprecated, please remove the call from your script.

Usually, you don't need to call \link{data_clean} explicitly,
it is called within plot, tab and idx functions.
}
\keyword{internal}
