% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{print.vlkr_rprt}
\alias{print.vlkr_rprt}
\title{Printing method for volker reports.}
\usage{
\method{print}{vlkr_rprt}(x, ...)
}
\arguments{
\item{x}{The volker report}

\item{...}{Further parameters passed to print}
}
\description{
Printing method for volker reports.
}
\keyword{internal}
