% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{print.vlkr_rprt}
\alias{print.vlkr_rprt}
\title{Printing method for volker reports}
\usage{
\method{print}{vlkr_rprt}(x, ...)
}
\arguments{
\item{x}{The volker report object.}

\item{...}{Further parameters passed to print.}
}
\value{
No return value.
}
\description{
Printing method for volker reports
}
\examples{
library(volker)
data <- volker::chatgpt

rp <- report_metrics(data, sd_age)
print(rp)

}
\keyword{internal}
