% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{print.vlkr_tbl}
\alias{print.vlkr_tbl}
\title{Printing method for volker tables.}
\usage{
\method{print}{vlkr_tbl}(obj)
}
\arguments{
\item{obj}{The volker table}
}
\description{
Printing method for volker tables.
}
