% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{print.vlkr_tbl}
\alias{print.vlkr_tbl}
\title{Printing method for volker tables.}
\usage{
\method{print}{vlkr_tbl}(x, ...)
}
\arguments{
\item{x}{The volker table.}

\item{...}{Further parameters passed to print().}
}
\value{
No return value.
}
\description{
Printing method for volker tables.
}
\examples{
library(volker)
data <- volker::chatgpt

tb <- tab_metrics(data, sd_age)
print(tb)

}
\keyword{internal}
