% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{relevance}
\alias{relevance}
\title{Calculate relevance of terms for a topic}
\usage{
relevance(p, lambda = 0.6)
}
\arguments{
\item{p}{The matrix of topic-term-probabilities.
Each row is a topic. Each column is a term.}
}
\value{
A matrix of the same shape as p with relevance scores
}
\description{
Relevance is p for lambda = 1 and the probability shift for lambda = 0.
}
\details{
See section 3 in https://nlp.stanford.edu/events/illvi2014/papers/sievert-illvi2014.pdf
See http://vis.stanford.edu/files/2012-Termite-AVI.pdf and
https://nlp.stanford.edu/events/illvi2014/papers/sievert-illvi2014.pdf
}
\author{
Inspired by Chantal Gaertner
}
