% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{replace_item_values}
\alias{replace_item_values}
\title{Replace items in the first column by labels found in the dataset}
\usage{
replace_item_values(result, data, cols)
}
\arguments{
\item{result}{The table with items in the first column,
e.g. produced by tab_item_count()}

\item{data}{The labeled dataset}

\item{cols}{The tidyselect columns}
}
\description{
Replace items in the first column by labels found in the dataset
}
