% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{report}
\alias{report}
\title{Automatically detect the table type and output table and plot}
\usage{
report(
  data,
  scopes,
  col_group = NULL,
  prop = "total",
  numbers = "p",
  missings = F,
  ordered = NULL,
  index = T,
  title = T,
  close = T
)
}
\arguments{
\item{data}{A tibble}

\item{scopes}{A list of column prefixes, e.g. "UM" or "WB" (character).}

\item{col_group}{A tidyselect column holding groups to compare, e.g. sd_gender (column name without quotes)}

\item{prop}{The basis of percent calculation when comparing groups: total (the default), cols, or rows}

\item{numbers}{The numbers plotted on bars: p (the default), n, both or NULL.}

\item{missings}{Include missing values (default FALSE)}

\item{ordered}{Values can be nominal (0) or ordered ascending (1) descending (-1).
By default (NULL), the ordering is automatically detected.
An appropriate color scale should be choosen depending on the ordering.
For unordered values, the default scale is used.
For ordered values, shades of the VLKR_FILLGRADIENT option are used.}

\item{index}{When multiple metric items are matched by a scope, an index will be calculated using the psych-package.
Set to FALSE to suppress index generation.}

\item{title}{A character providing the heading or TRUE (default) to output a heading.
Classes for tabset pills will be added.}

\item{close}{Whether to close the last tab (default value TRUE) or to keep it open.
Keep it open to add further custom tabs by adding headers on the fifth level
in Markdown (e.g. ##### Method)}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{report} has been deprecated. Use \link{report_counts}) and \link{report_metrics} instead.
}
\details{
For each column prefix:
\itemize{
\item If it matches a single column, use tab_counts_one and plot_counts_one
\item If it matches multiple columns, use tab_counts_items and plot_counts_items
}
}
\keyword{internal}
