% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Automatically detect the table type and output table and plot}
\usage{
report(data, scopes, col_group = NULL, prop = "total")
}
\arguments{
\item{data}{A tibble}

\item{scopes}{A list of column prefixes, e.g. "UM" or "WB" (character).}

\item{col_group}{A tidyselect column holding groups to compare, e.g. sd_gender (column name without quotes)}

\item{prop}{The basis of percent calculation when comparing groups: total (the default), cols, or rows}
}
\description{
For each column prefix:
\itemize{
\item If it matches a single column, use tab_var_counts and plot_var_counts
\item If it matches multiple columns, use tab_items_counts and plot_item_counts
}
}
\details{
TODO: Check whether the matched items are present
}
