% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Automatically detect the table type and output table and plot}
\usage{
report(
  data,
  scopes,
  col_group = NULL,
  prop = "total",
  numbers = "p",
  missings = F,
  ordered = NULL,
  index = T,
  title = T,
  close = T
)
}
\arguments{
\item{data}{A tibble}

\item{scopes}{A list of column prefixes, e.g. "UM" or "WB" (character).}

\item{col_group}{A tidyselect column holding groups to compare, e.g. sd_gender (column name without quotes)}

\item{prop}{The basis of percent calculation when comparing groups: total (the default), cols, or rows}

\item{numbers}{The numbers plotted on bars: p (the default), n, both or NULL.}

\item{missings}{Include missing values (default FALSE)}

\item{ordered}{Values can be nominal (0) or ordered ascending (1) descending (-1).
By default (NULL), the ordering is automatically detected.
An appropriate color scale should be choosen depending on the ordering.
For unordered values, the default scale is used.
For ordered values, the viridis scale is used.}

\item{index}{When multiple metric items are matched by a scope, an index will be calculated using the psych-package.
Set to FALSE to suppress index generation.}

\item{title}{A character providing the heading or TRUE (default) to output a heading.
Classes for tabset pills will be added.}

\item{close}{Whether to close the last tab (default value TRUE) or to keep it open.
Keep it open to add further custom tabs by adding headers on the fifth level
in Markdown (e.g. ##### Method)}
}
\description{
For each column prefix:
\itemize{
\item If it matches a single column, use tab_var_counts and plot_var_counts
\item If it matches multiple columns, use tab_items_counts and plot_item_counts
}
}
\details{
TODO: Check whether the matched items are present
}
