% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{report_styles}
\alias{report_styles}
\title{Pill navigation for switching between plots and tables in markdown reports}
\usage{
report_styles()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{report_styles} has been deprecated. Use the volker::html_document template instead.
}
\details{
Call report_styles() in one of your first code chunks to let the magic happen.
Make sure the code chunk's include option is set to TRUE.
Outputs a style tag with css styles that modify the standard bootstrap
theme.

See the html_report() format for an alternative to include the styles.
}
\keyword{internal}
