% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_styles}
\alias{report_styles}
\title{Pill navigation for switching between plots and tables in markdown reports}
\usage{
report_styles()
}
\description{
Call report_styles() in one of your first code chunks to let the magic happen.
Make sure the code chunk's include option is set to TRUE.
Outputs a style tag with css styles that modify the standard bootstrap
theme.
}
\details{
See the html_report() format for an alternative to include the styles.
}
