% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{set_col_label}
\alias{set_col_label}
\title{Alias for set_col_labels}
\usage{
set_col_label(data, cols, labels)
}
\description{
#' @description
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
\code{set_col_labels} has been deprecated. Use \link{labs_apply} instead.
}
\keyword{internal}
