% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{set_col_labels}
\alias{set_col_labels}
\alias{set_col_label}
\title{Set column labels by their comment attribute}
\usage{
set_col_labels(data, cols, labels)

set_col_label(data, cols, labels)
}
\arguments{
\item{data}{A data frame}

\item{cols}{Tidyselect column names (e. g. a single column name without quotes)}

\item{labels}{A character vector with the same length as the column selection, containing new labels}
}
\description{
TODO: RENAME TO set_title?
}
\examples{
ds <- volker::chatgpt
ds <- set_col_labels(ds, sd_alter, "Age")

}
