% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{set_col_labels}
\alias{set_col_labels}
\title{Set column labels by their comment attribute}
\usage{
set_col_labels(data, cols, labels)
}
\arguments{
\item{data}{A data frame}

\item{cols}{Tidyselect column names (e. g. a single column name without quotes)}

\item{labels}{A character vector with the same length as the column selection, containing new labels}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{set_col_labels} has been deprecated. Use \link{labs_apply} instead.
}
\examples{
ds <- volker::chatgpt
ds <- set_col_labels(ds, sd_age, "Age")

}
\keyword{internal}
