% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim.R
\name{skim_metrics}
\alias{skim_metrics}
\title{A reduced skimmer for metric variables
Returns a five point summary, mean and sd, items count and alpha for scales added by add_index()}
\usage{
skim_metrics(data, ..., .data_name = NULL)
}
\value{
A skimmer, see \link[skimr:skim_with]{skim_with}
}
\description{
A reduced skimmer for metric variables
Returns a five point summary, mean and sd, items count and alpha for scales added by add_index()
}
\examples{
library(volker)
data <- volker::chatgpt

skim_metrics(data)

}
\keyword{internal}
