% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stat_metrics_one_grouped}
\alias{stat_metrics_one_grouped}
\title{Output a regression table with estimates and makro statistics}
\usage{
stat_metrics_one_grouped(
  data,
  col,
  col_group,
  negative = FALSE,
  digits = 1,
  missings = FALSE,
  labels = TRUE,
  clean = TRUE,
  ...
)
}
\arguments{
\item{data}{A tibble}

\item{col}{The column holding metric values}

\item{col_group}{The column holding groups to compare}

\item{negative}{If FALSE (default), negative values are recoded as missing values.}

\item{digits}{The number of digits to print}

\item{missings}{Include missing values in the output (default FALSE)}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{stat_metrics}.}
}
\value{
A volker chunks object containing tables for micro and macro statistics
}
\description{
#TODO: Do we need the digits parameter?
#TODO: Add CI for effect sizes
}
\examples{
library(volker)
data <- volker::chatgpt

stat_metrics_one_grouped(data, sd_age, sd_gender)

}
\keyword{internal}
