% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_counts_items_cor}
\alias{tab_counts_items_cor}
\title{Correlate the values in multiple items}
\usage{
tab_counts_items_cor(
  data,
  cols1,
  cols2,
  values = c("n", "p"),
  category = NULL,
  digits = 1,
  .labels = T
)
}
\arguments{
\item{data}{A tibble}

\item{cols1}{The item columns}

\item{cols2}{The item columns}

\item{values}{The output values}

\item{category}{The category within the items to focus}

\item{digits}{The digits to print}

\item{.labels}{If True (default) extracts item labels from the attributes, see get_labels()}

\item{.negative}{If True (default), negative values are recoded to missing values}
}
\description{
TODO: implement -> calculate npmi or similar for one value
}
\keyword{internal}
