% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_counts_items_cor}
\alias{tab_counts_items_cor}
\title{Correlate the values in multiple items}
\usage{
tab_counts_items_cor(data, cols, cols_cor)
}
\arguments{
\item{data}{A tibble}

\item{cols}{The source columns}

\item{cols_cor}{The target columns or NULL to calculate correlations within the source columns}
}
\description{
TODO: implement -> calculate npmi or similar for one value
}
\keyword{internal}
