% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_counts_items_grouped}
\alias{tab_counts_items_grouped}
\title{Compare the values in multiple items by a grouping column}
\usage{
tab_counts_items_grouped(data, cols, cross, clean = TRUE, ...)
}
\arguments{
\item{data}{A tibble containing item measures.}

\item{cols}{The item columns that hold the values to summarize.}

\item{cross}{The column holding groups to compare.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{tab_counts}.}
}
\value{
A volker tibble.
}
\description{
\strong{Not yet implemented. The future will come.}
}
\keyword{internal}
