% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_counts_items_grouped}
\alias{tab_counts_items_grouped}
\title{Compare the values in multiple items by a grouping column.}
\usage{
tab_counts_items_grouped(data, cols, col_group, ...)
}
\arguments{
\item{data}{A tibble}

\item{cols}{The item columns that hold the values to summarize}

\item{col_group}{The column holding groups to compare}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{tab_counts}.}
}
\description{
TODO: implement -> focus one category and show n / p
}
\keyword{internal}
