% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_counts_one}
\alias{tab_counts_one}
\title{Output a frequency table for the values in one column}
\usage{
tab_counts_one(data, col, .labels = T, .formatted = T)
}
\arguments{
\item{data}{A tibble}

\item{col}{The column holding values to count}

\item{.labels}{If True (default) extracts item labels from the attributes, see get_labels()}

\item{.formatted}{Set to FALSE to prevent calculating percents from proportions}
}
\description{
Output a frequency table for the values in one column
}
