% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_counts_one}
\alias{tab_counts_one}
\title{Output a frequency table for the values in one column}
\usage{
tab_counts_one(
  data,
  col,
  ci = FALSE,
  percent = TRUE,
  labels = TRUE,
  clean = TRUE,
  ...
)
}
\arguments{
\item{data}{A tibble.}

\item{col}{The column holding values to count.}

\item{ci}{Whether to compute 95\% confidence intervals using \code{stats::\link[stats:prop.test]{prop.test}}.}

\item{percent}{Proportions are formatted as percent by default. Set to FALSE to get bare proportions.}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{tab_counts}.}
}
\value{
A volker tibble.
}
\description{
Output a frequency table for the values in one column
}
\examples{
library(volker)
data <- volker::chatgpt

tab_counts_one(data, sd_gender)

}
\keyword{internal}
