% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_group_counts}
\alias{tab_group_counts}
\title{Output frequencies cross tabulated with a grouping column}
\usage{
tab_group_counts(data, col, col_group, values = c("n", "p"), .quiet = F)
}
\arguments{
\item{data}{A tibble}

\item{col}{The column holding factor values}

\item{col_group}{The column holding groups to compare}

\item{values}{The values to output: n (frequency) or p (percentage).}

\item{.quiet}{Set to true to suppress printing the output}
}
\description{
Output frequencies cross tabulated with a grouping column
}
