% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_group_counts}
\alias{tab_group_counts}
\title{Output frequencies cross tabulated with a grouping column}
\usage{
tab_group_counts(
  data,
  col,
  col_group,
  values = c("n", "p"),
  prop = "total",
  .formatted = T,
  .labels = T
)
}
\arguments{
\item{data}{A tibble}

\item{col}{The column holding factor values}

\item{col_group}{The column holding groups to compare}

\item{values}{The values to output: n (frequency) or p (percentage).}

\item{prop}{The basis of percent calculation: total (the default), cols, or rows}

\item{.formatted}{Set to FALSE to prevent calculating percents from proportions}

\item{.labels}{If True (default) extracts item labels from the attributes, see get_labels()}
}
\description{
Output frequencies cross tabulated with a grouping column
}
