% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_group_metrics}
\alias{tab_group_metrics}
\title{#' Output a five point summary for groups}
\usage{
tab_group_metrics(data, col, col_group, digits = 1, .quiet = F)
}
\arguments{
\item{data}{A tibble}

\item{col}{The column holding metric values}

\item{col_group}{The column holding groups to compare}

\item{.quiet}{Set to true to suppress printing the output}
}
\description{
#' Output a five point summary for groups
}
