% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_group_metrics}
\alias{tab_group_metrics}
\title{#' Output a five point summary for groups}
\usage{
tab_group_metrics(data, col, col_group, digits = 1, .labels = T)
}
\arguments{
\item{data}{A tibble}

\item{col}{The column holding metric values}

\item{col_group}{The column holding groups to compare}

\item{digits}{The digits to print}

\item{.labels}{If True (default) extracts item labels from the attributes, see get_labels()}
}
\description{
#' Output a five point summary for groups
}
