% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_group_metrics}
\alias{tab_group_metrics}
\title{Output the means for groups in one or multiple columns}
\usage{
tab_group_metrics(
  data,
  cols_values,
  cols_groups,
  value = "numeric.mean",
  digits = 1,
  .quiet = F
)
}
\arguments{
\item{data}{A tibble}

\item{cols_values}{The columns that hold the values to summarize}

\item{value}{The output metric, choosen from values of skim_metrics()}

\item{digits}{The digits to print}

\item{.quiet}{Set to true to suppress printing the output}

\item{cols_group}{The columns holding groups to compare}
}
\description{
Output the means for groups in one or multiple columns
}
