% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_item_counts}
\alias{tab_item_counts}
\title{Output frequencies for multiple variables}
\usage{
tab_item_counts(data, cols_items, value = "n", .quiet = F)
}
\arguments{
\item{data}{A tibble containing item measures}

\item{cols_items}{Tidyselect item variables (e.g. starts_with...)}

\item{value}{The value to output: n (frequency) or p (percentage)}

\item{.quiet}{Set to true to suppress printing the output}
}
\description{
Output frequencies for multiple variables
}
