% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_item_counts}
\alias{tab_item_counts}
\title{Output frequencies for multiple variables}
\usage{
tab_item_counts(
  data,
  cols,
  values = c("n", "p"),
  missings = F,
  .formatted = T,
  .labels = T
)
}
\arguments{
\item{data}{A tibble containing item measures}

\item{cols}{Tidyselect item variables (e.g. starts_with...)}

\item{values}{The values to output: n (frequency) or p (percentage)}

\item{missings}{Include missing values (default FALSE)}

\item{.formatted}{Set to FALSE to prevent calculating percents from proportions}

\item{.labels}{If True (default) extracts item labels from the attributes, see get_labels()}

\item{.quiet}{Set to true to suppress printing the output}
}
\description{
TODO: Reorder boolean categories: first TRUE, then FALSE
TODO: Support single columns
}
