% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_item_metrics}
\alias{tab_item_metrics}
\title{Output a five point summary table for multiple items}
\usage{
tab_item_metrics(data, cols, digits = 1, .labels = T, .quiet = F)
}
\arguments{
\item{data}{A tibble}

\item{cols}{The columns holding metric values}

\item{digits}{The digits to print}

\item{.labels}{If True (default) extracts item labels from the attributes, see get_labels()}

\item{.quiet}{Set to true to suppress printing the output}
}
\description{
Output a five point summary table for multiple items
}
