% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_item_metrics}
\alias{tab_item_metrics}
\title{Output a five point summary table for groups in a variable}
\usage{
tab_item_metrics(data, cols_values, digits = 1, .quiet = F)
}
\arguments{
\item{data}{A tibble}

\item{cols_values}{The columns holding metric values}

\item{digits}{The digits to print}

\item{.quiet}{Set to true to suppress printing the output}
}
\description{
Output a five point summary table for groups in a variable
}
