% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_metrics_items}
\alias{tab_metrics_items}
\title{Output a five point summary table for multiple items}
\usage{
tab_metrics_items(data, cols, negative = F, digits = 1, labels = T, ...)
}
\arguments{
\item{data}{A tibble}

\item{cols}{The columns holding metric values}

\item{negative}{If FALSE (default), negative values are recoded as missing values.}

\item{digits}{The number of digits to print}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{tab_metrics}.}
}
\description{
Output a five point summary table for multiple items
}
