% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_metrics_items}
\alias{tab_metrics_items}
\title{Output a five point summary table for multiple items}
\usage{
tab_metrics_items(
  data,
  cols,
  ci = FALSE,
  digits = 1,
  labels = TRUE,
  clean = TRUE,
  ...
)
}
\arguments{
\item{data}{A tibble.}

\item{cols}{The columns holding metric values.}

\item{ci}{Whether to compute confidence intervals of the mean.}

\item{digits}{The number of digits to print.}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{tab_metrics}.}
}
\value{
A volker tibble.
}
\description{
Output a five point summary table for multiple items
}
\examples{
library(volker)
data <- volker::chatgpt

tab_metrics_items(data, starts_with("cg_adoption_"))

}
\keyword{internal}
