% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_metrics_items_cor}
\alias{tab_metrics_items_cor}
\title{Output a correlation table}
\usage{
tab_metrics_items_cor(
  data,
  cols,
  cols_cor,
  method = "p",
  significant = F,
  labels = T
)
}
\arguments{
\item{data}{A tibble}

\item{cols}{The source columns}

\item{cols_cor}{The target columns or NULL to calculate correlations within the source columns}

\item{method}{The output metrics, p = Pearson's R, s = Spearman's rho}

\item{significant}{Only show significant values}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
