% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_metrics_one}
\alias{tab_metrics_one}
\title{Output a five point summary table for the values in multiple columns}
\usage{
tab_metrics_one(
  data,
  col,
  ci = FALSE,
  digits = 1,
  labels = TRUE,
  clean = TRUE,
  ...
)
}
\arguments{
\item{data}{A tibble.}

\item{col}{The columns holding metric values.}

\item{ci}{Whether to calculate 95\% confidence intervals of the mean.}

\item{digits}{The number of digits to print.}

\item{labels}{If TRUE (default) extracts labels from the attributes, see \link{codebook}.}

\item{clean}{Prepare data by \link{data_clean}.}

\item{...}{Placeholder to allow calling the method with unused parameters from \link{tab_metrics}.}
}
\value{
A volker tibble.
}
\description{
Output a five point summary table for the values in multiple columns
}
\examples{
library(volker)
data <- volker::chatgpt

tab_metrics_one(data, sd_age)

}
\keyword{internal}
