% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{tab_multi_means}
\alias{tab_multi_means}
\title{Alias for tab_metrics_items_grouped}
\usage{
tab_multi_means(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{tab_multi_means} has been deprecated. Use \link{tab_metrics_items_grouped} instead.
}
\keyword{internal}
