% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_var_counts}
\alias{tab_var_counts}
\title{Output a frequency table for the values in one column}
\usage{
tab_var_counts(data, col, .quiet = F)
}
\arguments{
\item{data}{A tibble}

\item{col}{The column holding values to count}

\item{.quiet}{Set to true to suppress printing the output}
}
\description{
Output a frequency table for the values in one column
}
