% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_var_metrics}
\alias{tab_var_metrics}
\title{Output a five point summary table for the values multiple columns}
\usage{
tab_var_metrics(data, col, digits = 1, .quiet = F)
}
\arguments{
\item{data}{A tibble}

\item{col}{The columns holding metric values}

\item{digits}{The digits to print}

\item{.quiet}{Set to true to suppress printing the output}
}
\description{
Output a five point summary table for the values multiple columns
}
