% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{tab_var_metrics}
\alias{tab_var_metrics}
\title{Alias for tab_metrics_one}
\usage{
tab_var_metrics(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{tab_var_metrics} has been deprecated. Use \link{tab_metrics_one} instead.
}
\keyword{internal}
