% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{trunc_labels}
\alias{trunc_labels}
\title{Truncate labels}
\usage{
trunc_labels(x, max_length = 20)
}
\arguments{
\item{x}{A character vector.}

\item{max_length}{Maximum length, default is 20. The ellipsis "..." is appended to shortened labels.}
}
\value{
A character vector with truncated labels.
}
\description{
Truncate labels that exceed a specified maximum length.
}
\keyword{internal}
