% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{vlkr_colors_polarized}
\alias{vlkr_colors_polarized}
\title{Get colors for polarized scales}
\usage{
vlkr_colors_polarized(n = NULL)
}
\arguments{
\item{n}{Number of colors or NULL to get the raw colors from the config}
}
\value{
A vector of colors.
}
\description{
Creates a gradient scale based on VLKR_FILLPOLARIZED.
}
\keyword{internal}
