% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{vlkr_colors_sequential}
\alias{vlkr_colors_sequential}
\title{Get colors for sequential scales}
\usage{
vlkr_colors_sequential(n = NULL, inv = FALSE)
}
\arguments{
\item{n}{Number of colors or NULL to get the raw colors from the config}

\item{inv}{Whether to get a text color with good contrast on the chosen fill colors.}
}
\value{
A vector of colors.
}
\description{
Creates a gradient scale based on VLKR_FILLGRADIENT.
}
\keyword{internal}
