% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{VLKR_NA_NUMERIC}
\alias{VLKR_NA_NUMERIC}
\title{Numbers to remove from vectors}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
VLKR_NA_NUMERIC
}
\description{
By default the value \code{-9} is removed  from tables and plots
}
\keyword{internal}
