% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{VLKR_PLOT_LABELWRAP}
\alias{VLKR_PLOT_LABELWRAP}
\title{Wrapping threshold}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
VLKR_PLOT_LABELWRAP
}
\description{
Override with \code{options(vlkr.wrap.labels=20)}.
Override with \code{options(vlkr.wrap.legend=10)}.
Override with \code{options(vlkr.wrap.scale=10)}.
}
\keyword{internal}
