% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volker-package.R
\docType{package}
\name{volker-package}
\alias{volker}
\alias{volker-package}
\title{volker: Report summaries and graphics for survey data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The package provides functions to create tables and plots to be included in Markdown reports. Boiled down to the basics, with labeling features and automatic output selection.
}
\author{
\strong{Maintainer}: Jakob Jünger \email{jakob.juenger@uni-muenster.de}

Other contributors:
\itemize{
  \item Henrieke Kotthoff \email{henrieke.kotthoff@uni-muenster.de} [contributor]
}

}
\keyword{internal}
