% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volker-package.R
\docType{package}
\name{volker-package}
\alias{volker}
\alias{volker-package}
\title{volker: A Package for Survey Data Summaries and Graphics}
\description{
This package contains functions for creating summaries and graphics for survey data.
}
\author{
\strong{Maintainer}: Jakob Jünger \email{jakob.juenger@uni-muenster.de}

Other contributors:
\itemize{
  \item Henrieke Kotthoff \email{henrieke.kotthoff@uni-muenster.de} [contributor]
}

}
\keyword{internal}
